@@ -293,13 +293,13 @@ impl Display for Message {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
             Message::Ping(ping) => {
-                write!(f, "ping tx={:?}", &ping.tx_id[..6])
+                write!(f, "Ping(tx={})", hex::encode(ping.tx_id))
             }
             Message::Pong(pong) => {
-                write!(f, "ping tx={:?}", &pong.tx_id[..6])
+                write!(f, "Pong(tx={})", hex::encode(pong.tx_id))
             }
             Message::CallMeMaybe(_) => {
-                write!(f, "call-me-maybe")
+                write!(f, "CallMeMaybe")
             }
         }
     }
