@@ -6,6 +6,8 @@ use std::{
     fmt::{Debug, Display},
     hash::Hash,
     str::FromStr,
+    sync::Mutex,
+    time::Duration,
 };
 
 pub use ed25519_dalek::{Signature, PUBLIC_KEY_LENGTH};
@@ -14,22 +16,84 @@ use once_cell::sync::OnceCell;
 use rand_core::CryptoRngCore;
 use serde::{Deserialize, Serialize};
 use ssh_key::LineEnding;
+use ttl_cache::TtlCache;
 
 pub use self::encryption::SharedSecret;
 use self::encryption::{public_ed_box, secret_ed_box};
 
-/// A public key.
-#[derive(Clone, Copy, PartialEq, Eq)]
-pub struct PublicKey {
-    public: VerifyingKey,
-    /// Cached version of `crypto_box::PublicKey` matching `public`.
-    /// Stored as raw array, as `crypto_box::PublicKey` is not `Copy`.
-    public_crypto_box: [u8; 32],
+#[derive(Debug)]
+struct CryptoKeys {
+    verifying_key: VerifyingKey,
+    crypto_box: crypto_box::PublicKey,
 }
+
+impl CryptoKeys {
+    fn new(verifying_key: VerifyingKey) -> Self {
+        let crypto_box = public_ed_box(&verifying_key);
+        Self {
+            verifying_key,
+            crypto_box,
+        }
+    }
+}
+
+/// Expiry time for the crypto key cache.
+///
+/// Basically, if no crypto operations have been performed with a key for this
+/// duration, the crypto keys will be removed from the cache and need to be
+/// re-created when they are used again.
+const KEY_CACHE_TTL: Duration = Duration::from_secs(60);
+/// Maximum number of keys in the crypto key cache. CryptoKeys are 224 bytes,
+/// keys are 32 bytes, so each entry is 256 bytes plus some overhead.
+///
+/// So that is about 4MB of max memory for the cache.
+const KEY_CACHE_CAPACITY: usize = 1024 * 16;
+static KEY_CACHE: OnceCell<Mutex<TtlCache<[u8; 32], CryptoKeys>>> = OnceCell::new();
+
+fn lock_key_cache() -> std::sync::MutexGuard<'static, TtlCache<[u8; 32], CryptoKeys>> {
+    let mutex = KEY_CACHE.get_or_init(|| Mutex::new(TtlCache::new(KEY_CACHE_CAPACITY)));
+    mutex.lock().unwrap()
+}
+
+/// Get or create the crypto keys, and project something out of them.
+///
+/// If the key has been verified before, this will not fail.
+fn get_or_create_crypto_keys<T>(
+    key: &[u8; 32],
+    f: impl Fn(&CryptoKeys) -> T,
+) -> std::result::Result<T, SignatureError> {
+    let mut state = lock_key_cache();
+    Ok(match state.entry(*key) {
+        ttl_cache::Entry::Occupied(entry) => {
+            // cache hit
+            f(entry.get())
+        }
+        ttl_cache::Entry::Vacant(entry) => {
+            // cache miss, create. This might fail if the key is invalid.
+            let vk = VerifyingKey::from_bytes(key)?;
+            let item = CryptoKeys::new(vk);
+            let item = entry.insert(item, KEY_CACHE_TTL);
+            f(item)
+        }
+    })
+}
+
+/// A public key.
+///
+/// The key itself is just a 32 byte array, but a key has associated crypto
+/// information that is cached for performance reasons.
+///
+/// The cache item will be refreshed every time a crypto operation is performed,
+/// or when a key is deserialised or created from a byte array.
+///
+/// Serialisation or creation from a byte array is cheap if the key is already
+/// in the cache, but expensive if it is not.
+#[derive(Clone, Copy, PartialEq, Eq)]
+pub struct PublicKey([u8; 32]);
 
 impl Hash for PublicKey {
     fn hash<H: std::hash::Hasher>(&self, state: &mut H) {
-        self.public.hash(state);
+        self.0.hash(state);
     }
 }
 
@@ -38,7 +102,7 @@ impl Serialize for PublicKey {
     where
         S: serde::Serializer,
     {
-        self.public.serialize(serializer)
+        serializer.serialize_bytes(&self.0)
     }
 }
 
@@ -47,15 +111,24 @@ impl<'de> Deserialize<'de> for PublicKey {
     where
         D: serde::Deserializer<'de>,
     {
-        let public = VerifyingKey::deserialize(deserializer)?;
-        Ok(public.into())
+        let bytes: &serde_bytes::Bytes = serde::Deserialize::deserialize(deserializer)?;
+        Self::try_from(bytes.as_ref()).map_err(serde::de::Error::custom)
     }
 }
 
 impl PublicKey {
     /// Get this public key as a byte array.
     pub fn as_bytes(&self) -> &[u8; 32] {
-        self.public.as_bytes()
+        &self.0
+    }
+
+    fn public(&self) -> VerifyingKey {
+        get_or_create_crypto_keys(&self.0, |item| item.verifying_key).expect("key has been checked")
+    }
+
+    fn public_crypto_box(&self) -> crypto_box::PublicKey {
+        get_or_create_crypto_keys(&self.0, |item| item.crypto_box.clone())
+            .expect("key has been checked")
     }
 
     /// Construct a `PublicKey` from a slice of bytes.
@@ -66,12 +139,8 @@ impl PublicKey {
     /// a valid `ed25519_dalek` curve point. Will never fail for bytes return from [`Self::as_bytes`].
     /// See [`VerifyingKey::from_bytes`] for details.
     pub fn from_bytes(bytes: &[u8; 32]) -> Result<Self, SignatureError> {
-        let public = VerifyingKey::from_bytes(bytes)?;
-        Ok(public.into())
-    }
-
-    fn public_crypto_box(&self) -> crypto_box::PublicKey {
-        crypto_box::PublicKey::from_bytes(self.public_crypto_box)
+        get_or_create_crypto_keys(bytes, |item| item.verifying_key)?;
+        Ok(Self(*bytes))
     }
 
     /// Verify a signature on a message with this secret key's public key.
@@ -80,7 +149,7 @@ impl PublicKey {
     ///
     /// Returns `Ok(())` if the signature is valid, and `Err` otherwise.
     pub fn verify(&self, message: &[u8], signature: &Signature) -> Result<(), SignatureError> {
-        self.public.verify_strict(message, signature)
+        self.public().verify_strict(message, signature)
     }
 
     /// Convert to a base32 string limited to the first 10 bytes for a friendly string
@@ -97,8 +166,21 @@ impl TryFrom<&[u8]> for PublicKey {
 
     #[inline]
     fn try_from(bytes: &[u8]) -> Result<Self, Self::Error> {
-        let public = VerifyingKey::try_from(bytes)?;
-        Ok(public.into())
+        Ok(match <[u8; 32]>::try_from(bytes) {
+            Ok(bytes) => {
+                // using from_bytes is faster than going via the verifying
+                // key in case the key is already in the cache, which should
+                // be quite common.
+                Self::from_bytes(&bytes)?
+            }
+            Err(_) => {
+                // this will always fail since the size is wrong.
+                // but there is no public constructor for SignatureError,
+                // so ¯\_(ツ)_/¯...
+                let vk = VerifyingKey::try_from(bytes)?;
+                vk.into()
+            }
+        })
     }
 }
 
@@ -118,12 +200,14 @@ impl AsRef<[u8]> for PublicKey {
 }
 
 impl From<VerifyingKey> for PublicKey {
-    fn from(public: VerifyingKey) -> Self {
-        let public_crypto_box = public_ed_box(&public).to_bytes();
-        PublicKey {
-            public,
-            public_crypto_box,
-        }
+    fn from(verifying_key: VerifyingKey) -> Self {
+        let item = CryptoKeys::new(verifying_key);
+        let key = *verifying_key.as_bytes();
+        let mut table = lock_key_cache();
+        // we already have performed the crypto operation, so no need for
+        // get_or_create_crypto_keys. Just insert in any case.
+        table.insert(key, item, KEY_CACHE_TTL);
+        PublicKey(key)
     }
 }
 
@@ -336,5 +420,36 @@ mod tests {
             PublicKey::from_str(&key.public().to_string()).unwrap(),
             key.public()
         );
+    }
+
+    /// Test the different ways a key can come into existence, and that they
+    /// all populate the key cache.
+    #[test]
+    fn test_key_creation_cache() {
+        let random_verifying_key = || {
+            let sk = SigningKey::generate(&mut rand::thread_rng());
+            sk.verifying_key()
+        };
+        let random_public_key = || random_verifying_key().to_bytes();
+        let k1 = random_public_key();
+        let _key = PublicKey::from_bytes(&k1).unwrap();
+        assert!(lock_key_cache().contains_key(&k1));
+
+        let k2 = random_public_key();
+        let _key = PublicKey::try_from(&k2).unwrap();
+        assert!(lock_key_cache().contains_key(&k2));
+
+        let k3 = random_public_key();
+        let _key = PublicKey::try_from(k3.as_slice()).unwrap();
+        assert!(lock_key_cache().contains_key(&k3));
+
+        let k4 = random_verifying_key();
+        let _key = PublicKey::from(k4);
+        assert!(lock_key_cache().contains_key(k4.as_bytes()));
+
+        let k5 = random_verifying_key();
+        let bytes = postcard::to_stdvec(&k5).unwrap();
+        let _key: PublicKey = postcard::from_bytes(&bytes).unwrap();
+        assert!(lock_key_cache().contains_key(k5.as_bytes()));
     }
 }
