@@ -20,7 +20,7 @@
 
 use std::{
     collections::HashMap,
-    fmt::{Debug, Display},
+    fmt::Display,
     io,
     net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr},
     path::PathBuf,
@@ -28,16 +28,17 @@ use std::{
         atomic::{AtomicBool, AtomicU16, AtomicU64, Ordering},
         Arc,
     },
-    task::{Context, Poll, Waker},
+    task::{ready, Context, Poll, Waker},
     time::{Duration, Instant},
 };
 
-use anyhow::{bail, Context as _, Result};
+use anyhow::{Context as _, Result};
 use bytes::Bytes;
 use futures::{future::BoxFuture, FutureExt};
 use iroh_metrics::{inc, inc_by};
 use quinn::AsyncUdpSocket;
 use rand::{seq::SliceRandom, Rng, SeedableRng};
+use smallvec::{smallvec, SmallVec};
 use tokio::{
     sync::{self, mpsc, Mutex},
     time,
@@ -62,7 +63,6 @@ use self::{
     endpoint::{Options as EndpointOptions, PeerMap, PingAction},
     metrics::Metrics as MagicsockMetrics,
     rebinding_conn::RebindingUdpConn,
-    udp_actor::{IpPacket, NetworkReadResult, NetworkSource, UdpActor, UdpActorMessage},
 };
 
 mod derp_actor;
@@ -70,10 +70,9 @@ mod endpoint;
 mod metrics;
 mod rebinding_conn;
 mod timer;
-mod udp_actor;
 
 pub use self::endpoint::ConnectionType;
-pub use self::endpoint::EndpointInfo;
+pub use self::endpoint::{DirectAddrInfo, EndpointInfo};
 pub use self::metrics::Metrics;
 pub use self::timer::Timer;
 
@@ -152,12 +151,12 @@ pub struct Options {
     /// Path to store known peers.
     pub peers_path: Option<std::path::PathBuf>,
 
-    /// Optionally provides a discovery mechanism for finding peers.
+    /// Optional discovery mechanism.
     pub discovery: Option<Box<dyn Discovery>>,
 }
 
 /// Discovery trait for [`MagicEndpoint`].
-pub trait Discovery: Debug + Send + Sync {
+pub trait Discovery: std::fmt::Debug + Send + Sync {
     /// Publish the given [`AddrInfo`] to the discovery mechanisms.
     fn publish(&self, info: &AddrInfo);
 
@@ -189,11 +188,15 @@ impl Default for Options {
             secret_key: SecretKey::generate(),
             derp_map: DerpMap::empty(),
             callbacks: Default::default(),
-            discovery: None,
             peers_path: None,
+            discovery: None,
         }
     }
 }
+
+/// Contents of a DERP message. Use a SmallVec to avoid allocations for the very
+/// common case of a single packet.
+pub(crate) type DerpContents = SmallVec<[Bytes; 1]>;
 
 /// Iroh connectivity layer.
 ///
@@ -216,8 +219,7 @@ pub struct MagicSock {
 #[derive(derive_more::Debug)]
 struct Inner {
     actor_sender: mpsc::Sender<ActorMessage>,
-    /// Sends network messages.
-    network_sender: mpsc::Sender<Vec<quinn_udp::Transmit>>,
+    derp_actor_sender: mpsc::Sender<DerpActorMessage>,
     /// String representation of the peer_id of this node.
     me: String,
     #[allow(clippy::type_complexity)]
@@ -230,7 +232,7 @@ struct Inner {
     on_net_info: Option<Box<dyn Fn(config::NetInfo) + Send + Sync + 'static>>,
 
     /// Used for receiving DERP messages.
-    network_recv_ch: flume::Receiver<NetworkReadResult>,
+    derp_recv_receiver: flume::Receiver<DerpRecvResult>,
     /// Stores wakers, to be called when derp_recv_ch receives new data.
     network_recv_wakers: std::sync::Mutex<Option<Waker>>,
     network_send_wakers: std::sync::Mutex<Option<Waker>>,
@@ -255,8 +257,23 @@ struct Inner {
     derp_map: DerpMap,
     /// Nearest DERP region ID; 0 means none/unknown.
     my_derp: AtomicU16,
+    /// Tracks the networkmap node entity for each peer discovery key.
+    peer_map: PeerMap,
+    /// UDP IPv4 socket
+    pconn4: RebindingUdpConn,
+    /// UDP IPv6 socket
+    pconn6: Option<RebindingUdpConn>,
+    /// Netcheck client
+    net_checker: netcheck::Client,
+    /// The state for an active DiscoKey.
+    disco_info: parking_lot::Mutex<HashMap<PublicKey, DiscoInfo>>,
+    udp_state: quinn_udp::UdpState,
 
-    /// Optional discovery mechanism for finding peers.
+    // Send buffer used in `poll_send_udp`
+    send_buffer: parking_lot::Mutex<Vec<quinn_udp::Transmit>>,
+    // UDP disco (ping) queue
+    udp_disco_sender: mpsc::Sender<(SocketAddr, PublicKey, disco::Message)>,
+
     discovery: Option<Box<dyn Discovery>>,
 }
 
@@ -294,6 +311,762 @@ impl Inner {
 
     fn public_key(&self) -> PublicKey {
         self.secret_key.public()
+    }
+
+    /// Get the cached version of the Ipv4 and Ipv6 addrs of the current connection.
+    fn local_addr(&self) -> (SocketAddr, Option<SocketAddr>) {
+        *self.local_addrs.read().unwrap()
+    }
+    fn normalized_local_addr(&self) -> io::Result<SocketAddr> {
+        let (v4, v6) = self.local_addr();
+        let addr = if let Some(v6) = v6 { v6 } else { v4 };
+        Ok(addr)
+    }
+
+    #[instrument(skip_all, fields(me = %self.me))]
+    fn poll_send(
+        &self,
+        cx: &mut Context,
+        transmits: &[quinn_udp::Transmit],
+    ) -> Poll<io::Result<usize>> {
+        let bytes_total: usize = transmits.iter().map(|t| t.contents.len()).sum();
+        inc_by!(MagicsockMetrics, send_data, bytes_total as _);
+
+        if self.is_closed() {
+            inc_by!(MagicsockMetrics, send_data_network_down, bytes_total as _);
+            return Poll::Ready(Err(io::Error::new(
+                io::ErrorKind::NotConnected,
+                "connection closed",
+            )));
+        }
+
+        let mut n = 0;
+        if transmits.is_empty() {
+            return Poll::Ready(Ok(n));
+        }
+
+        trace!(
+            "sending:\n{}",
+            transmits.iter().fold(
+                String::with_capacity(transmits.len() * 50),
+                |mut final_repr, t| {
+                    final_repr.push_str(
+                        format!(
+                            "  dest: {}, src: {:?}, content_len: {}\n",
+                            QuicMappedAddr(t.destination),
+                            t.src_ip,
+                            t.contents.len()
+                        )
+                        .as_str(),
+                    );
+                    final_repr
+                }
+            )
+        );
+
+        let dest = transmits[0].destination;
+        for transmit in transmits.iter() {
+            if transmit.destination != dest {
+                break;
+            }
+            n += 1;
+        }
+
+        // Copy the transmits into an owned buffer, because we will have to modify the send
+        // addresses to translate from the quic mapped address to the actual UDP address.
+        // To avoid allocating on each call to `poll_send`, we use a fixed buffer.
+        let mut transmits = {
+            let mut buf = self.send_buffer.lock();
+            buf.clear();
+            buf.reserve(n);
+            buf.extend_from_slice(&transmits[..n]);
+            buf
+        };
+
+        let dest = QuicMappedAddr(dest);
+
+        match self.peer_map.get_send_addrs_for_quic_mapped_addr(&dest) {
+            Some((public_key, udp_addr, derp_region, mut msgs)) => {
+                let mut pings_sent = false;
+                // If we have pings to send, we *have* to send them out first.
+                if !msgs.is_empty() {
+                    if let Err(err) = ready!(self.poll_handle_ping_actions(cx, &mut msgs)) {
+                        warn!(peer = %public_key.fmt_short(), "failed to handle ping actions: {err:?}");
+                    }
+                    pings_sent = true;
+                }
+
+                let mut udp_sent = false;
+                let mut derp_sent = false;
+                let mut udp_error = None;
+
+                trace!(
+                    peer = %public_key.fmt_short(),
+                    quic_addr = %dest.0,
+                    transmit_count = %transmits.len(),
+                    packet_count = &transmits.iter().map(|t| t.segment_size.map(|ss| t.contents.len() / ss).unwrap_or(1)).sum::<usize>(),
+                    len = &transmits.iter().map(|t| t.contents.len()).sum::<usize>(),
+                    ?udp_addr,
+                    ?derp_region,
+                    "send transmits"
+                );
+
+                // send udp
+                if let Some(addr) = udp_addr {
+                    // rewrite target addresses.
+                    for t in transmits.iter_mut() {
+                        t.destination = addr;
+                    }
+                    match ready!(self.poll_send_udp(addr, &transmits, cx)) {
+                        Ok(n) => {
+                            debug!(peer = %public_key.fmt_short(), dst = %addr, transmit_count=n, "sent transmits over UDP");
+                            // truncate the transmits vec to `n`. these transmits will be sent to
+                            // Derp further below. We only want to send those transmits to Derp that were
+                            // sent to UDP, because the next transmits will be sent on the next
+                            // call to poll_send, which will happen immediately after, because we
+                            // are always returning Poll::Ready if poll_send_udp returned
+                            // Poll::Ready.
+                            transmits.truncate(n);
+                            udp_sent = true;
+                            // record metrics.
+                        }
+                        Err(err) => {
+                            error!(peer = %public_key.fmt_short(), ?addr, "failed to send udp: {err:?}");
+                            udp_error = Some(err);
+                        }
+                    }
+                }
+
+                let n = transmits.len();
+
+                // send derp
+                if let Some(derp_region) = derp_region {
+                    self.send_derp(derp_region, public_key, split_packets(&transmits));
+                    derp_sent = true;
+                }
+
+                if !derp_sent && !udp_sent && !pings_sent {
+                    warn!(peer = %public_key.fmt_short(), "failed to send: no UDP or DERP addr");
+                    let err = udp_error.unwrap_or_else(|| {
+                        io::Error::new(
+                            io::ErrorKind::NotConnected,
+                            "no UDP or Derp address available for peer",
+                        )
+                    });
+                    Poll::Ready(Err(err))
+                } else {
+                    Poll::Ready(Ok(n))
+                }
+            }
+            None => {
+                error!(dst=%dest, "no endpoint for mapped address");
+                Poll::Ready(Err(io::Error::new(
+                    io::ErrorKind::NotConnected,
+                    "trying to send to unknown endpoint",
+                )))
+            }
+        }
+    }
+
+    fn poll_send_udp(
+        &self,
+        addr: SocketAddr,
+        transmits: &[quinn_udp::Transmit],
+        cx: &mut Context<'_>,
+    ) -> Poll<io::Result<usize>> {
+        let conn = self.conn_for_addr(addr)?;
+        let n = ready!(conn.poll_send(&self.udp_state, cx, transmits))?;
+        let total_bytes: u64 = transmits
+            .iter()
+            .take(n)
+            .map(|x| x.contents.len() as u64)
+            .sum();
+        if addr.is_ipv6() {
+            inc_by!(MagicsockMetrics, send_ipv6, total_bytes);
+        } else {
+            inc_by!(MagicsockMetrics, send_ipv4, total_bytes);
+        }
+        Poll::Ready(Ok(n))
+    }
+
+    fn conn_for_addr(&self, addr: SocketAddr) -> io::Result<&RebindingUdpConn> {
+        if addr.is_ipv6() && self.pconn6.is_none() {
+            return Err(io::Error::new(io::ErrorKind::Other, "no IPv6 connection"));
+        }
+        Ok(if addr.is_ipv6() {
+            self.pconn6.as_ref().unwrap()
+        } else {
+            &self.pconn4
+        })
+    }
+
+    #[instrument(skip_all, fields(me = %self.me))]
+    fn poll_recv(
+        &self,
+        cx: &mut Context,
+        bufs: &mut [io::IoSliceMut<'_>],
+        metas: &mut [quinn_udp::RecvMeta],
+    ) -> Poll<io::Result<usize>> {
+        // FIXME: currently ipv4 load results in ipv6 traffic being ignored
+        debug_assert_eq!(bufs.len(), metas.len(), "non matching bufs & metas");
+        if self.is_closed() {
+            return Poll::Ready(Err(io::Error::new(
+                io::ErrorKind::NotConnected,
+                "connection closed",
+            )));
+        }
+
+        // order of polling is: UDPv4, UDPv6, Derp
+        let msgs = match self.pconn4.poll_recv(cx, bufs, metas)? {
+            Poll::Pending | Poll::Ready(0) => match &self.pconn6 {
+                Some(conn) => match conn.poll_recv(cx, bufs, metas)? {
+                    Poll::Pending | Poll::Ready(0) => {
+                        return self.poll_recv_derp(cx, bufs, metas);
+                    }
+                    Poll::Ready(n) => n,
+                },
+                None => {
+                    return self.poll_recv_derp(cx, bufs, metas);
+                }
+            },
+            Poll::Ready(n) => n,
+        };
+
+        let dst_ip = self.normalized_local_addr().ok().map(|addr| addr.ip());
+
+        let mut quic_packets_total = 0;
+
+        for (meta, buf) in metas.iter_mut().zip(bufs.iter_mut()).take(msgs) {
+            let mut start = 0;
+            let mut is_quic = false;
+            let mut quic_packets_count = 0;
+
+            // find disco and stun packets and forward them to the actor
+            loop {
+                let end = start + meta.stride;
+                if end > meta.len {
+                    break;
+                }
+                let packet = &buf[start..end];
+                let packet_is_quic = if stun::is(packet) {
+                    trace!(src = %meta.addr, len = %meta.stride, "UDP recv: stun packet");
+                    let packet2 = Bytes::copy_from_slice(packet);
+                    self.net_checker.receive_stun_packet(packet2, meta.addr);
+                    false
+                } else if let Some((sender, sealed_box)) = disco::source_and_box(packet) {
+                    // Disco?
+                    trace!(src = %meta.addr, len = %meta.stride, "UDP recv: disco packet");
+                    self.handle_disco_message(
+                        sender,
+                        sealed_box,
+                        DiscoMessageSource::Udp(meta.addr),
+                    );
+                    false
+                } else {
+                    trace!(src = %meta.addr, len = %meta.stride, "UDP recv: quic packet");
+                    true
+                };
+
+                if packet_is_quic {
+                    quic_packets_count += 1;
+                    is_quic = true;
+                } else {
+                    // overwrite the first byte of the packets with zero.
+                    // this makes quinn reliably and quickly ignore the packet as long as
+                    // [`quinn::EndpointConfig::grease_quic_bit`] is set to `true`.
+                    buf[start] = 0u8;
+                }
+                start = end;
+            }
+
+            if is_quic {
+                // remap addr
+                match self.peer_map.get_quic_mapped_addr_for_ip_port(meta.addr) {
+                    None => {
+                        warn!(src = ?meta.addr, count = %quic_packets_count, len = meta.len, "recv packets: no peer state found, skipping");
+                        // if we have no peer state for the from addr, set len to 0 to make quinn skip the buf completely.
+                        meta.len = 0;
+                    }
+                    Some(quic_mapped_addr) => {
+                        trace!(src = ?meta.addr, count = %quic_packets_count, len = meta.len, "recv packets: peer state found");
+                        quic_packets_total += quic_packets_count;
+                        meta.addr = quic_mapped_addr.0;
+                    }
+                }
+            } else {
+                // if there is no non-stun,non-disco packet in the chunk, set len to zero to make
+                // quinn skip the buf completely.
+                meta.len = 0;
+            }
+            // Normalize local_ip
+            meta.dst_ip = dst_ip;
+        }
+
+        if quic_packets_total > 0 {
+            inc_by!(MagicsockMetrics, recv_datagrams, quic_packets_total as _);
+            trace!("UDP recv: {} packets", quic_packets_total);
+        }
+
+        Poll::Ready(Ok(msgs))
+    }
+
+    #[instrument(skip_all, fields(name = %self.me))]
+    fn poll_recv_derp(
+        &self,
+        cx: &mut Context,
+        bufs: &mut [io::IoSliceMut<'_>],
+        metas: &mut [quinn_udp::RecvMeta],
+    ) -> Poll<io::Result<usize>> {
+        let mut num_msgs = 0;
+        for (buf_out, meta_out) in bufs.iter_mut().zip(metas.iter_mut()) {
+            if self.is_closed() {
+                break;
+            }
+            match self.derp_recv_receiver.try_recv() {
+                Err(flume::TryRecvError::Empty) => {
+                    self.network_recv_wakers
+                        .lock()
+                        .unwrap()
+                        .replace(cx.waker().clone());
+                    break;
+                }
+                Err(flume::TryRecvError::Disconnected) => {
+                    return Poll::Ready(Err(io::Error::new(
+                        io::ErrorKind::NotConnected,
+                        "connection closed",
+                    )));
+                }
+                Ok(Err(err)) => return Poll::Ready(Err(err)),
+                Ok(Ok((meta, bytes))) => {
+                    inc_by!(MagicsockMetrics, recv_data_derp, bytes.len() as _);
+                    trace!(src = %meta.addr, len = %meta.len, "recv packet from derp");
+                    buf_out[..bytes.len()].copy_from_slice(&bytes);
+                    *meta_out = meta;
+                    num_msgs += 1;
+                }
+            }
+        }
+
+        // If we have any msgs to report, they are in the first `num_msgs_total` slots
+        if num_msgs > 0 {
+            inc_by!(MagicsockMetrics, recv_datagrams, num_msgs as _);
+            trace!("derp recv: {} packets", num_msgs);
+            return Poll::Ready(Ok(num_msgs));
+        }
+
+        Poll::Pending
+    }
+
+    /// Handles a discovery message.
+    #[instrument("disco_in", skip_all, fields(peer = %sender.fmt_short(), %src))]
+    fn handle_disco_message(&self, sender: PublicKey, sealed_box: &[u8], src: DiscoMessageSource) {
+        trace!("handle_disco_message start");
+        if self.is_closed() {
+            return;
+        }
+
+        // We're now reasonably sure we're expecting communication from
+        // this peer, do the heavy crypto lifting to see what they want.
+        //
+        let (payload, sealed_box) = {
+            let mut disco_info = self.disco_info.lock();
+            let di = get_disco_info(&mut disco_info, &self.secret_key, &sender);
+            let mut sealed_box = sealed_box.to_vec();
+            let payload = di.shared_key.open(&mut sealed_box);
+            (payload, sealed_box)
+        };
+
+        if payload.is_err() {
+            // This could happen if we changed the key between restarts.
+            warn!("disco: failed to open box (wrong rcpt?) {:?}", payload,);
+            inc!(MagicsockMetrics, recv_disco_bad_key);
+            return;
+        }
+        let dm = disco::Message::from_bytes(&sealed_box);
+
+        let dm = match dm {
+            Ok(dm) => dm,
+            Err(err) => {
+                // Couldn't parse it, but it was inside a correctly
+                // signed box, so just ignore it, assuming it's from a
+                // newer version of Tailscale that we don't
+                // understand. Not even worth logging about, lest it
+                // be too spammy for old clients.
+
+                inc!(MagicsockMetrics, recv_disco_bad_parse);
+                debug!(?err, "failed to parse disco message");
+                return;
+            }
+        };
+
+        let is_derp = src.is_derp();
+        if is_derp {
+            inc!(MagicsockMetrics, recv_disco_derp);
+        } else {
+            inc!(MagicsockMetrics, recv_disco_udp);
+        }
+
+        trace!(message = ?dm, "receive disco message");
+        match dm {
+            disco::Message::Ping(ping) => {
+                inc!(MagicsockMetrics, recv_disco_ping);
+                self.handle_ping(ping, &sender, src);
+            }
+            disco::Message::Pong(pong) => {
+                inc!(MagicsockMetrics, recv_disco_pong);
+                self.peer_map.write(|pm| {
+                    if let Some(ep) = pm.endpoint_for_node_key_mut(&sender).as_mut() {
+                        let insert = ep.handle_pong_conn(&pong, src.into());
+                        if let Some((src, key)) = insert {
+                            pm.set_node_key_for_ip_port(src, &key);
+                        }
+                        debug!(?insert, "received pong")
+                    } else {
+                        warn!("received pong: peer unknown, ignore")
+                    }
+                });
+            }
+            disco::Message::CallMeMaybe(cm) => {
+                inc!(MagicsockMetrics, recv_disco_call_me_maybe);
+                let DiscoMessageSource::Derp { key, .. } = src else {
+                    // CallMeMaybe messages should only come via DERP.
+                    debug!("[unexpected] call-me-maybe packets should only come via DERP");
+                    return;
+                };
+                self.peer_map
+                    .write(|pm| match pm.endpoint_for_node_key_mut(&key).as_mut() {
+                        None => {
+                            inc!(MagicsockMetrics, recv_disco_call_me_maybe_bad_disco);
+                            debug!("received call-me-maybe: ignore, peer is unknown");
+                        }
+                        Some(ep) => {
+                            debug!("received call-me-maybe: {} endpoints", cm.my_number.len());
+                            ep.handle_call_me_maybe(cm);
+                        }
+                    });
+            }
+        }
+    }
+
+    /// Handle a ping message.
+    fn handle_ping(&self, dm: disco::Ping, sender: &PublicKey, src: DiscoMessageSource) {
+        let src_addr: SendAddr = src.clone().into();
+        let (node_key, likely_heart_beat) = {
+            let mut disco_info = self.disco_info.lock();
+            let di = get_disco_info(&mut disco_info, &self.secret_key, sender);
+            let likely_heart_beat = Some(src_addr) == di.last_ping_from
+                && di
+                    .last_ping_time
+                    .map(|s| s.elapsed() < Duration::from_secs(5))
+                    .unwrap_or_default();
+            di.last_ping_from.replace(src_addr);
+            di.last_ping_time.replace(Instant::now());
+            (di.node_key, likely_heart_beat)
+        };
+        debug_assert_eq!(sender, &node_key);
+
+        let dst_key = match src {
+            DiscoMessageSource::Derp { key, .. } => key,
+            DiscoMessageSource::Udp(_) => node_key,
+        };
+
+        if !likely_heart_beat {
+            debug!(tx = %hex::encode(dm.tx_id), "received ping");
+        }
+
+        let unknown_sender = self.peer_map.write(|pm| {
+            if pm.endpoint_for_node_key(sender).is_none() {
+                match src {
+                    DiscoMessageSource::Udp(addr) => pm.receive_ip(addr).is_none(),
+                    DiscoMessageSource::Derp { .. } => true,
+                }
+            } else {
+                false
+            }
+        });
+
+        // if we get here we got a valid ping from an unknown sender
+        // so insert an endpoint for them
+        if unknown_sender {
+            debug!("received ping: peer unknown, add to peer map");
+            self.peer_map.write(|pm| {
+                pm.insert_endpoint(EndpointOptions {
+                    public_key: *sender,
+                    derp_region: src.derp_region(),
+                    active: true,
+                })
+            });
+        }
+
+        // Insert the ping into the peer map, and return whether a ping with this tx_id was already
+        // received.
+        let is_duplicate = self.peer_map.write(|pm| {
+            if let Some(ep) = pm.endpoint_for_node_key_mut(&dst_key) {
+                if ep.endpoint_confirmed(src_addr, dm.tx_id) {
+                    return true;
+                }
+
+                if let DiscoMessageSource::Udp(addr) = src {
+                    pm.set_node_key_for_ip_port(addr, &dst_key);
+                }
+            }
+            false
+        });
+
+        if is_duplicate {
+            debug!(%src, tx = %hex::encode(dm.tx_id), "received ping: endpoint already confirmed, skip");
+            return;
+        };
+
+        // Send a pong.
+        debug!(tx = %hex::encode(dm.tx_id), "send pong");
+        let pong = disco::Message::Pong(disco::Pong {
+            tx_id: dm.tx_id,
+            src: src.as_socket_addr(),
+        });
+        match src {
+            DiscoMessageSource::Udp(addr) => {
+                if let Err(err) = self.udp_disco_sender.try_send((addr, dst_key, pong)) {
+                    warn!(%addr, %err, "failed to queue pong");
+                }
+            }
+            DiscoMessageSource::Derp { region, .. } => {
+                self.send_disco_message_derp(region, dst_key, pong);
+            }
+        }
+    }
+
+    fn encode_disco_message(&self, dst_key: PublicKey, msg: &disco::Message) -> Bytes {
+        let seal = {
+            let mut disco_info = self.disco_info.lock();
+            let di = get_disco_info(&mut disco_info, &self.secret_key, &dst_key);
+            let mut seal = msg.as_bytes();
+            di.shared_key.seal(&mut seal);
+            seal
+        };
+
+        disco::encode_message(&self.public_key(), seal).into()
+    }
+
+    fn send_disco_message_derp(&self, region: u16, dst_key: PublicKey, msg: disco::Message) {
+        trace!(peer = %dst_key.fmt_short(), %region, %msg, "send disco message (derp)");
+        let pkt = self.encode_disco_message(dst_key, &msg);
+        inc!(MagicsockMetrics, send_disco_derp);
+        self.send_derp(region, dst_key, smallvec![pkt]);
+        inc!(MagicsockMetrics, sent_disco_derp);
+        disco_message_sent(&msg);
+    }
+
+    async fn send_disco_message_udp(
+        &self,
+        dst: SocketAddr,
+        dst_key: PublicKey,
+        msg: &disco::Message,
+    ) -> io::Result<bool> {
+        futures::future::poll_fn(move |cx| self.poll_send_disco_message_udp(dst, dst_key, msg, cx))
+            .await
+    }
+
+    fn poll_send_disco_message_udp(
+        &self,
+        dst: SocketAddr,
+        dst_key: PublicKey,
+        msg: &disco::Message,
+        cx: &mut Context<'_>,
+    ) -> Poll<std::io::Result<bool>> {
+        trace!(%dst, %msg, "send disco message (UDP)");
+        if self.is_closed() {
+            return Poll::Ready(Err(io::Error::new(
+                io::ErrorKind::NotConnected,
+                "connection closed",
+            )));
+        }
+        let pkt = self.encode_disco_message(dst_key, msg);
+        // TODO: These metrics will be wrong with the poll impl
+        // Also - do we need it? I'd say the `sent_disco_udp` below is enough.
+        inc!(MagicsockMetrics, send_disco_udp);
+        let transmits = [quinn_udp::Transmit {
+            destination: dst,
+            contents: pkt,
+            ecn: None,
+            segment_size: None,
+            src_ip: None, // TODO
+        }];
+        let sent = ready!(self.poll_send_udp(dst, &transmits, cx));
+        Poll::Ready(match sent {
+            Ok(0) => {
+                // Can't send. (e.g. no IPv6 locally)
+                warn!(?msg, "failed to send disco message");
+                Ok(false)
+            }
+            Ok(_n) => {
+                debug!(%msg, "sent disco message");
+                inc!(MagicsockMetrics, sent_disco_udp);
+                disco_message_sent(msg);
+                Ok(true)
+            }
+            Err(err) => {
+                warn!(?msg, ?err, "failed to send disco message");
+                Err(err)
+            }
+        })
+    }
+
+    fn poll_handle_ping_actions(
+        &self,
+        cx: &mut Context<'_>,
+        msgs: &mut Vec<PingAction>,
+    ) -> Poll<io::Result<()>> {
+        if msgs.is_empty() {
+            return Poll::Ready(Ok(()));
+        }
+
+        while let Some(msg) = msgs.pop() {
+            if self.poll_handle_ping_action(cx, &msg)?.is_pending() {
+                msgs.push(msg);
+                return Poll::Pending;
+            }
+        }
+        Poll::Ready(Ok(()))
+    }
+
+    #[instrument("handle_ping_action", skip_all)]
+    fn poll_handle_ping_action(
+        &self,
+        cx: &mut Context<'_>,
+        msg: &PingAction,
+    ) -> Poll<io::Result<()>> {
+        // Abort sending as soon as we know we are shutting down.
+        if self.is_closing() || self.is_closed() {
+            return Poll::Ready(Ok(()));
+        }
+        match *msg {
+            PingAction::EnqueueCallMeMaybe {
+                derp_region,
+                endpoint_id,
+            } => {
+                let msg = ActorMessage::EnqueueCallMeMaybe {
+                    derp_region,
+                    endpoint_id,
+                };
+                match self.actor_sender.try_send(msg) {
+                    Ok(_) => {}
+                    Err(mpsc::error::TrySendError::Closed(_)) => {
+                        error!("magicsock actor dropped");
+                    }
+                    Err(mpsc::error::TrySendError::Full(_)) => {
+                        warn!("actor channel is full, dropping CallMeMaybe message");
+                    }
+                }
+            }
+            PingAction::SendPing {
+                id,
+                dst,
+                dst_key,
+                tx_id,
+                purpose,
+            } => {
+                let msg = disco::Message::Ping(disco::Ping {
+                    tx_id,
+                    node_key: self.public_key(),
+                });
+                match dst {
+                    SendAddr::Udp(addr) => {
+                        ready!(self.poll_send_disco_message_udp(addr, dst_key, &msg, cx))?;
+                    }
+                    SendAddr::Derp(region) => self.send_disco_message_derp(region, dst_key, msg),
+                }
+                let msg_sender = self.actor_sender.clone();
+                self.peer_map.write(move |pm| {
+                    if let Some(ep) = pm.by_id_mut(&id) {
+                        ep.ping_sent(dst, tx_id, purpose, msg_sender);
+                    }
+                });
+            }
+        }
+        Poll::Ready(Ok(()))
+    }
+
+    fn send_derp(&self, region_id: u16, peer: PublicKey, contents: DerpContents) {
+        trace!(peer = %peer.fmt_short(), derp_region = region_id, count = contents.len(), len = contents.iter().map(|c| c.len()).sum::<usize>(), "send derp");
+        let msg = DerpActorMessage::Send {
+            region_id,
+            contents,
+            peer,
+        };
+        match self.derp_actor_sender.try_send(msg) {
+            Ok(_) => {
+                trace!(peer = %peer.fmt_short(), derp_region = region_id, "send derp: message queued")
+            }
+            Err(mpsc::error::TrySendError::Closed(_)) => {
+                warn!(peer = %peer.fmt_short(), derp_region = region_id, "send derp: message dropped, channel to actor is closed");
+            }
+            Err(mpsc::error::TrySendError::Full(_)) => {
+                warn!(peer = %peer.fmt_short(), derp_region = region_id, "send derp: message dropped, channel to actor is full");
+            }
+        }
+    }
+
+    #[instrument(skip_all)]
+    async fn send_raw(
+        &self,
+        addr: SocketAddr,
+        mut transmits: Vec<quinn_udp::Transmit>,
+    ) -> io::Result<usize> {
+        trace!(dst = %addr, "send {} packets", transmits.len());
+
+        if transmits.iter().any(|t| t.destination != addr) {
+            for t in &mut transmits {
+                t.destination = addr;
+            }
+        }
+
+        futures::future::poll_fn(|cx| self.poll_send_udp(addr, &transmits, cx)).await
+    }
+}
+
+#[derive(Clone, Debug)]
+enum DiscoMessageSource {
+    Udp(SocketAddr),
+    Derp { region: u16, key: PublicKey },
+}
+
+impl Display for DiscoMessageSource {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            Self::Udp(addr) => write!(f, "Udp({addr})"),
+            Self::Derp { region, key } => write!(f, "Derp({region}, {})", key.fmt_short()),
+        }
+    }
+}
+
+impl From<DiscoMessageSource> for SendAddr {
+    fn from(value: DiscoMessageSource) -> Self {
+        match value {
+            DiscoMessageSource::Udp(addr) => SendAddr::Udp(addr),
+            DiscoMessageSource::Derp { region, .. } => SendAddr::Derp(region),
+        }
+    }
+}
+
+impl DiscoMessageSource {
+    fn is_derp(&self) -> bool {
+        matches!(self, DiscoMessageSource::Derp { .. })
+    }
+
+    fn derp_region(&self) -> Option<u16> {
+        match self {
+            Self::Derp { region, .. } => Some(*region),
+            Self::Udp(_) => None,
+        }
+    }
+
+    /// Returns the mapped version or the actual `SocketAddr`.
+    fn as_socket_addr(&self) -> SocketAddr {
+        match self {
+            Self::Derp { region, .. } => SocketAddr::new(DERP_MAGIC_IP, *region),
+            Self::Udp(addr) => *addr,
+        }
     }
 }
 
@@ -371,7 +1144,7 @@ impl MagicSock {
             None => None,
         };
 
-        let (network_recv_ch_sender, network_recv_ch_receiver) = flume::bounded(128);
+        let (derp_recv_sender, derp_recv_receiver) = flume::bounded(128);
 
         let (pconn4, pconn6) = bind(port).await?;
         let port = pconn4.port();
@@ -387,56 +1160,10 @@ impl MagicSock {
         let ipv6_addr = pconn6.as_ref().and_then(|c| c.local_addr().ok());
 
         let net_checker = netcheck::Client::new(Some(port_mapper.clone())).await?;
+
         let (actor_sender, actor_receiver) = mpsc::channel(128);
-        let (network_sender, network_receiver) = mpsc::channel(128);
-
-        let inner = Arc::new(Inner {
-            me,
-            on_endpoints,
-            on_derp_active,
-            on_net_info,
-            port: AtomicU16::new(port),
-            secret_key,
-            local_addrs: std::sync::RwLock::new((ipv4_addr, ipv6_addr)),
-            closing: AtomicBool::new(false),
-            closed: AtomicBool::new(false),
-            network_recv_ch: network_recv_ch_receiver,
-            network_recv_wakers: std::sync::Mutex::new(None),
-            network_send_wakers: std::sync::Mutex::new(None),
-            actor_sender: actor_sender.clone(),
-            network_sender,
-            ipv6_reported: Arc::new(AtomicBool::new(false)),
-            derp_map,
-            my_derp: AtomicU16::new(0),
-            discovery,
-        });
-
-        let udp_state = quinn_udp::UdpState::default();
-        let (ip_sender, ip_receiver) = mpsc::channel(128);
-        let (udp_actor_sender, udp_actor_receiver) = mpsc::channel(128);
-
-        let udp_actor_task = {
-            let udp_actor =
-                UdpActor::new(&udp_state, inner.clone(), pconn4.clone(), pconn6.clone());
-            let net_checker = net_checker.clone();
-            tokio::task::spawn(
-                async move {
-                    udp_actor
-                        .run(udp_actor_receiver, net_checker, ip_sender)
-                        .await;
-                }
-                .instrument(info_span!("udp.actor")),
-            )
-        };
-
         let (derp_actor_sender, derp_actor_receiver) = mpsc::channel(256);
-        let derp_actor = DerpActor::new(inner.clone(), actor_sender.clone());
-        let derp_actor_task = tokio::task::spawn(
-            async move {
-                derp_actor.run(derp_actor_receiver).await;
-            }
-            .instrument(info_span!("derp.actor")),
-        );
+        let (udp_disco_sender, mut udp_disco_receiver) = mpsc::channel(256);
 
         // load the peer data
         let peer_map = match peers_path.as_ref() {
@@ -454,6 +1181,54 @@ impl MagicSock {
             _ => PeerMap::default(),
         };
 
+        let udp_state = quinn_udp::UdpState::default();
+
+        let inner = Arc::new(Inner {
+            me,
+            on_endpoints,
+            on_derp_active,
+            on_net_info,
+            port: AtomicU16::new(port),
+            secret_key,
+            local_addrs: std::sync::RwLock::new((ipv4_addr, ipv6_addr)),
+            closing: AtomicBool::new(false),
+            closed: AtomicBool::new(false),
+            derp_recv_receiver,
+            network_recv_wakers: std::sync::Mutex::new(None),
+            network_send_wakers: std::sync::Mutex::new(None),
+            actor_sender: actor_sender.clone(),
+            ipv6_reported: Arc::new(AtomicBool::new(false)),
+            derp_map,
+            my_derp: AtomicU16::new(0),
+            pconn4: pconn4.clone(),
+            pconn6: pconn6.clone(),
+            net_checker: net_checker.clone(),
+            disco_info: parking_lot::Mutex::new(HashMap::new()),
+            peer_map,
+            derp_actor_sender: derp_actor_sender.clone(),
+            udp_state,
+            send_buffer: Default::default(),
+            udp_disco_sender,
+            discovery,
+        });
+
+        let derp_actor = DerpActor::new(inner.clone(), actor_sender.clone());
+        let derp_actor_task = tokio::task::spawn(
+            async move {
+                derp_actor.run(derp_actor_receiver).await;
+            }
+            .instrument(info_span!("derp.actor")),
+        );
+
+        let inner2 = inner.clone();
+        let udp_disco_sender_task = tokio::task::spawn(async move {
+            while let Some((dst, dst_key, msg)) = udp_disco_receiver.recv().await {
+                if let Err(err) = inner2.send_disco_message_udp(dst, dst_key, &msg).await {
+                    warn!(%dst, peer = %dst_key.fmt_short(), ?err, "failed to send disco message (UDP)");
+                }
+            }
+        });
+
         let inner2 = inner.clone();
         let main_actor_task = tokio::task::spawn(
             async move {
@@ -461,26 +1236,21 @@ impl MagicSock {
                     msg_receiver: actor_receiver,
                     msg_sender: actor_sender,
                     derp_actor_sender,
-                    udp_actor_sender,
-                    network_receiver,
-                    ip_receiver,
                     inner: inner2,
-                    derp_recv_sender: network_recv_ch_sender,
+                    derp_recv_sender,
                     endpoints_update_state: EndpointUpdateState::new(),
                     last_endpoints: Vec::new(),
                     last_endpoints_time: None,
                     on_endpoint_refreshed: HashMap::new(),
                     periodic_re_stun_timer: new_re_stun_timer(false),
                     net_info_last: None,
-                    disco_info: HashMap::new(),
-                    peer_map,
                     peers_path,
                     port_mapper,
                     pconn4,
                     pconn6,
-                    udp_state,
                     no_v4_send: false,
                     net_checker,
+                    udp_disco_sender_task,
                 };
 
                 if let Err(err) = actor.run().await {
@@ -495,7 +1265,6 @@ impl MagicSock {
             actor_tasks: Arc::new(Mutex::new(vec![
                 main_actor_task.into(),
                 derp_actor_task.into(),
-                udp_actor_task.into(),
             ])),
         };
 
@@ -536,7 +1305,7 @@ impl MagicSock {
 
     /// Get the cached version of the Ipv4 and Ipv6 addrs of the current connection.
     pub fn local_addr(&self) -> Result<(SocketAddr, Option<SocketAddr>)> {
-        Ok(*self.inner.local_addrs.read().unwrap())
+        Ok(self.inner.local_addr())
     }
 
     /// Triggers an address discovery. The provided why string is for debug logging only.
@@ -628,14 +1397,8 @@ impl MagicSock {
 
     #[instrument(skip_all, fields(me = %self.inner.me))]
     /// Add addresses for a node to the magic socket's addresbook.
-    pub async fn add_peer_addr(&self, addr: PeerAddr) -> Result<()> {
-        let (s, r) = sync::oneshot::channel();
-        self.inner
-            .actor_sender
-            .send(ActorMessage::AddKnownAddr(addr, s))
-            .await?;
-        r.await?;
-        Ok(())
+    pub fn add_peer_addr(&self, addr: PeerAddr) {
+        self.inner.peer_map.add_peer_addr(addr);
     }
 
     /// Closes the connection.
@@ -675,8 +1438,9 @@ impl MagicSock {
         r.await.unwrap();
     }
 
-    pub(crate) fn discovery(&self) -> &Option<Box<dyn Discovery>> {
-        &self.inner.discovery
+    /// Reference to optional discovery service
+    pub fn discovery(&self) -> Option<&dyn Discovery> {
+        self.inner.discovery.as_ref().map(Box::as_ref)
     }
 }
 
@@ -686,6 +1450,7 @@ impl MagicSock {
 /// node. In the case of shared nodes and users switching accounts, two
 /// nodes in the NetMap may legitimately have the same DiscoKey.  As
 /// such, no fields in here should be considered node-specific.
+#[derive(Debug)]
 struct DiscoInfo {
     node_key: PublicKey,
     /// The precomputed key for communication with the peer that has the `node_key` used to
@@ -699,6 +1464,8 @@ struct DiscoInfo {
     /// The last time of a ping for `node_key`.
     last_ping_time: Option<Instant>,
 }
+
+type DerpRecvResult = Result<(quinn_udp::RecvMeta, Bytes), io::Error>;
 
 /// Reports whether x and y represent the same set of endpoints. The order doesn't matter.
 fn endpoint_sets_equal(xs: &[config::Endpoint], ys: &[config::Endpoint]) -> bool {
@@ -729,146 +1496,22 @@ fn endpoint_sets_equal(xs: &[config::Endpoint], ys: &[config::Endpoint]) -> bool
 }
 
 impl AsyncUdpSocket for MagicSock {
-    #[instrument(skip_all, fields(me = %self.inner.me))]
     fn poll_send(
         &self,
         _udp_state: &quinn_udp::UdpState,
         cx: &mut Context,
         transmits: &[quinn_udp::Transmit],
     ) -> Poll<io::Result<usize>> {
-        let bytes_total: usize = transmits.iter().map(|t| t.contents.len()).sum();
-        inc_by!(MagicsockMetrics, send_data, bytes_total as _);
-
-        if self.inner.is_closed() {
-            inc_by!(MagicsockMetrics, send_data_network_down, bytes_total as _);
-            return Poll::Ready(Err(io::Error::new(
-                io::ErrorKind::NotConnected,
-                "connection closed",
-            )));
-        }
-
-        let mut n = 0;
-        if transmits.is_empty() {
-            return Poll::Ready(Ok(n));
-        }
-
-        // Split up transmits by destination, as the rest of the code assumes single dest.
-        let groups = TransmitIter::new(transmits);
-        for group in groups {
-            match self.inner.network_sender.try_reserve() {
-                Err(mpsc::error::TrySendError::Full(_)) => {
-                    // TODO: add counter?
-                    self.inner
-                        .network_send_wakers
-                        .lock()
-                        .unwrap()
-                        .replace(cx.waker().clone());
-                    break;
-                }
-                Err(mpsc::error::TrySendError::Closed(_)) => {
-                    return Poll::Ready(Err(io::Error::new(
-                        io::ErrorKind::NotConnected,
-                        "connection closed",
-                    )));
-                }
-                Ok(permit) => {
-                    n += group.len();
-                    permit.send(group);
-                }
-            }
-        }
-        if n > 0 {
-            return Poll::Ready(Ok(n));
-        }
-
-        Poll::Pending
+        self.inner.poll_send(cx, transmits)
     }
 
-    #[instrument(skip_all, fields(me = %self.inner.me))]
     fn poll_recv(
         &self,
         cx: &mut Context,
         bufs: &mut [io::IoSliceMut<'_>],
         metas: &mut [quinn_udp::RecvMeta],
     ) -> Poll<io::Result<usize>> {
-        // FIXME: currently ipv4 load results in ipv6 traffic being ignored
-        debug_assert_eq!(bufs.len(), metas.len(), "non matching bufs & metas");
-        if self.inner.is_closed() {
-            return Poll::Ready(Err(io::Error::new(
-                io::ErrorKind::NotConnected,
-                "connection closed",
-            )));
-        }
-
-        let mut num_msgs = 0;
-        for (buf_out, meta_out) in bufs.iter_mut().zip(metas.iter_mut()) {
-            match self.inner.network_recv_ch.try_recv() {
-                Err(flume::TryRecvError::Empty) => {
-                    self.inner
-                        .network_recv_wakers
-                        .lock()
-                        .unwrap()
-                        .replace(cx.waker().clone());
-                    break;
-                }
-                Err(flume::TryRecvError::Disconnected) => {
-                    return Poll::Ready(Err(io::Error::new(
-                        io::ErrorKind::NotConnected,
-                        "connection closed",
-                    )));
-                }
-                Ok(dm) => {
-                    if self.inner.is_closed() {
-                        break;
-                    }
-
-                    match dm {
-                        NetworkReadResult::Error(err) => {
-                            return Poll::Ready(Err(err));
-                        }
-                        NetworkReadResult::Ok {
-                            bytes,
-                            meta,
-                            source,
-                        } => {
-                            buf_out[..bytes.len()].copy_from_slice(&bytes);
-                            *meta_out = meta;
-
-                            match source {
-                                NetworkSource::Derp => {
-                                    inc_by!(MagicsockMetrics, recv_data_derp, bytes.len() as _);
-                                }
-                                NetworkSource::Ipv4 => {
-                                    inc_by!(MagicsockMetrics, recv_data_ipv4, bytes.len() as _);
-                                }
-                                NetworkSource::Ipv6 => {
-                                    inc_by!(MagicsockMetrics, recv_data_ipv6, bytes.len() as _);
-                                }
-                            }
-                            trace!(
-                                "[QUINN] <- {} ({}b) ({}) ({:?}, {:?})",
-                                meta_out.addr,
-                                meta_out.len,
-                                self.inner.me,
-                                meta_out.dst_ip,
-                                source
-                            );
-                        }
-                    }
-
-                    num_msgs += 1;
-                }
-            }
-        }
-
-        // If we have any msgs to report, they are in the first `num_msgs_total` slots
-        if num_msgs > 0 {
-            inc_by!(MagicsockMetrics, recv_datagrams, num_msgs as _);
-            trace!("received {} datagrams", num_msgs);
-            return Poll::Ready(Ok(num_msgs));
-        }
-
-        Poll::Pending
+        self.inner.poll_recv(cx, bufs, metas)
     }
 
     fn local_addr(&self) -> io::Result<SocketAddr> {
@@ -888,7 +1531,6 @@ impl AsyncUdpSocket for MagicSock {
 }
 
 #[derive(Debug)]
-#[allow(clippy::large_enum_variant)]
 enum ActorMessage {
     TrackedEndpoints(sync::oneshot::Sender<Vec<EndpointInfo>>),
     TrackedEndpoint(PublicKey, sync::oneshot::Sender<Option<EndpointInfo>>),
@@ -897,21 +1539,13 @@ enum ActorMessage {
     SetPreferredPort(u16, sync::oneshot::Sender<()>),
     RebindAll(sync::oneshot::Sender<()>),
     Shutdown,
-    CloseOrReconnect(u16, &'static str),
     ReStun(&'static str),
     EnqueueCallMeMaybe {
         derp_region: u16,
         endpoint_id: usize,
     },
-    SendCallMeMaybe {
-        dst: SendAddr,
-        dst_key: PublicKey,
-        msg: disco::CallMeMaybe,
-    },
-    AddKnownAddr(PeerAddr, sync::oneshot::Sender<()>),
     ReceiveDerp(DerpReadResult),
     EndpointPingExpired(usize, stun::TransactionId),
-    PublishAddrInfo,
 }
 
 struct Actor {
@@ -919,11 +1553,8 @@ struct Actor {
     msg_receiver: mpsc::Receiver<ActorMessage>,
     msg_sender: mpsc::Sender<ActorMessage>,
     derp_actor_sender: mpsc::Sender<DerpActorMessage>,
-    udp_actor_sender: mpsc::Sender<UdpActorMessage>,
-    network_receiver: mpsc::Receiver<Vec<quinn_udp::Transmit>>,
-    ip_receiver: mpsc::Receiver<IpPacket>,
     /// Channel to send received derp messages on, for processing.
-    derp_recv_sender: flume::Sender<NetworkReadResult>,
+    derp_recv_sender: flume::Sender<DerpRecvResult>,
     /// Indicates the update endpoint state.
     endpoints_update_state: EndpointUpdateState,
     /// Records the endpoints found during the previous
@@ -940,17 +1571,12 @@ struct Actor {
     periodic_re_stun_timer: time::Interval,
     /// The `NetInfo` provided in the last call to `net_info_func`. It's used to deduplicate calls to netInfoFunc.
     net_info_last: Option<config::NetInfo>,
-    /// The state for an active DiscoKey.
-    disco_info: HashMap<PublicKey, DiscoInfo>,
-    /// Tracks the networkmap node entity for each peer discovery key.
-    peer_map: PeerMap,
-    /// Path where connection info from [`Self::peer_map`] is persisted.
+    /// Path where connection info from [`Inner::peer_map`] is persisted.
     peers_path: Option<PathBuf>,
 
     // The underlying UDP sockets used to send/rcv packets.
     pconn4: RebindingUdpConn,
     pconn6: Option<RebindingUdpConn>,
-    udp_state: quinn_udp::UdpState,
 
     /// The NAT-PMP/PCP/UPnP prober/client, for requesting port mappings from NAT devices.
     port_mapper: portmapper::Client,
@@ -962,6 +1588,9 @@ struct Actor {
 
     /// The prober that discovers local network conditions, including the closest DERP relay and NAT mappings.
     net_checker: netcheck::Client,
+
+    /// Task that sends disco messages over UDP.
+    udp_disco_sender_task: tokio::task::JoinHandle<()>,
 }
 
 impl Actor {
@@ -999,7 +1628,7 @@ impl Actor {
 
         // Let the the hearbeat only start a couple seconds later
         let mut endpoint_heartbeat_timer = time::interval_at(
-            time::Instant::now() + Duration::from_secs(5),
+            time::Instant::now() + HEARTBEAT_INTERVAL,
             HEARTBEAT_INTERVAL,
         );
         let mut endpoints_update_receiver = self.endpoints_update_state.running.subscribe();
@@ -1015,35 +1644,10 @@ impl Actor {
 
         loop {
             tokio::select! {
-                Some(transmits) = self.network_receiver.recv() => {
-                    trace!("tick: network send");
-                    self.send_network(transmits).await;
-                }
                 Some(msg) = self.msg_receiver.recv() => {
                     trace!(?msg, "tick: msg");
                     if self.handle_actor_message(msg).await {
                         return Ok(());
-                    }
-                }
-                Some(msg) = self.ip_receiver.recv() => {
-                    trace!("tick: ip_receiver");
-                    match msg {
-                        IpPacket::Disco { sender, sealed_box, src } => {
-                            self.handle_disco_message(sender, &sealed_box, src, None).await;
-                        }
-                        IpPacket::Forward(mut forward) => {
-                            if let NetworkReadResult::Ok { meta, bytes, .. } = &mut forward {
-                                if !self.receive_ip(bytes, meta) {
-                                    continue;
-                                }
-                            }
-
-                            let _ = self.derp_recv_sender.send_async(forward).await;
-                            let mut wakers = self.inner.network_recv_wakers.lock().unwrap();
-                            while let Some(waker) = wakers.take() {
-                                waker.wake();
-                            }
-                        }
                     }
                 }
                 tick = self.periodic_re_stun_timer.tick() => {
@@ -1057,13 +1661,11 @@ impl Actor {
                     self.re_stun("portmap_updated").await;
                 },
                 _ = endpoint_heartbeat_timer.tick() => {
-                    trace!("tick: endpoint heartbeat {} endpoints", self.peer_map.node_count());
+                    trace!("tick: endpoint heartbeat {} endpoints", self.inner.peer_map.node_count());
                     // TODO: this might trigger too many packets at once, pace this
-                    self.peer_map.prune_inactive();
-                    let mut msgs = Vec::new();
-                    for (_, ep) in self.peer_map.endpoints_mut() {
-                        msgs.extend(ep.stayin_alive());
-                    }
+
+                    self.inner.peer_map.prune_inactive();
+                    let msgs = self.inner.peer_map.endpoints_stayin_alive();
                     self.handle_ping_actions(msgs).await;
                 }
                 _ = endpoints_update_receiver.changed() => {
@@ -1075,8 +1677,9 @@ impl Actor {
                 }
                 _ = save_peers_timer.tick(), if self.peers_path.is_some() => {
                     let path = self.peers_path.as_ref().expect("precondition: `is_some()`");
-                    self.peer_map.prune_inactive();
-                    match self.peer_map.save_to_file(path).await {
+
+                    self.inner.peer_map.prune_inactive();
+                    match self.inner.peer_map.save_to_file(path).await {
                         Ok(count) => debug!(count, "peers persisted"),
                         Err(e) => debug!(%e, "failed to persist known peers"),
                     }
@@ -1088,47 +1691,14 @@ impl Actor {
         }
     }
 
-    async fn handle_ping_actions(&mut self, msgs: Vec<PingAction>) {
+    async fn handle_ping_actions(&mut self, mut msgs: Vec<PingAction>) {
         if msgs.is_empty() {
             return;
         }
-
-        info!("handle_ping_actions ({}): {:?}", msgs.len(), msgs);
-        for msg in msgs {
-            // Abort sending as soon as we know we are shutting down.
-            if self.inner.is_closing() || self.inner.is_closed() {
-                break;
-            }
-            match msg {
-                PingAction::EnqueueCallMeMaybe {
-                    derp_region,
-                    endpoint_id,
-                } => {
-                    self.enqueue_call_me_maybe(derp_region, endpoint_id).await;
-                }
-                PingAction::SendPing {
-                    id,
-                    dst,
-                    dst_key,
-                    tx_id,
-                    purpose,
-                } => {
-                    let msg = disco::Message::Ping(disco::Ping {
-                        tx_id,
-                        node_key: self.inner.public_key(),
-                    });
-                    match self.send_disco_message(dst, dst_key, msg).await {
-                        Ok(true) => {
-                            if let Some(ep) = self.peer_map.by_id_mut(&id) {
-                                ep.ping_sent(dst, tx_id, purpose, self.msg_sender.clone());
-                            }
-                        }
-                        _ => {
-                            debug!("failed to send ping to {:?}", dst);
-                        }
-                    }
-                }
-            }
+        if let Err(err) =
+            futures::future::poll_fn(|cx| self.inner.poll_handle_ping_actions(cx, &mut msgs)).await
+        {
+            debug!("failed to send pings: {err:?}");
         }
     }
 
@@ -1138,11 +1708,11 @@ impl Actor {
     async fn handle_actor_message(&mut self, msg: ActorMessage) -> bool {
         match msg {
             ActorMessage::TrackedEndpoints(s) => {
-                let eps: Vec<_> = self.peer_map.endpoint_infos();
+                let eps: Vec<_> = self.inner.peer_map.endpoint_infos(Instant::now());
                 let _ = s.send(eps);
             }
             ActorMessage::TrackedEndpoint(node_key, s) => {
-                let _ = s.send(self.peer_map.endpoint_info(&node_key));
+                let _ = s.send(self.inner.peer_map.endpoint_info(&node_key));
             }
             ActorMessage::LocalEndpoints(s) => {
                 let eps: Vec<_> = self.last_endpoints.clone();
@@ -1150,18 +1720,17 @@ impl Actor {
             }
             ActorMessage::GetMappingAddr(node_key, s) => {
                 let res = self
+                    .inner
                     .peer_map
-                    .endpoint_for_node_key(&node_key)
-                    .map(|ep| ep.quic_mapped_addr);
+                    .get_quic_mapped_addr_for_node_key(&node_key);
                 let _ = s.send(res);
             }
             ActorMessage::Shutdown => {
                 debug!("shutting down");
-                for (_, ep) in self.peer_map.endpoints_mut() {
-                    ep.stop_and_reset();
-                }
+                self.udp_disco_sender_task.abort();
+                self.inner.peer_map.notify_shutdown();
                 if let Some(path) = self.peers_path.as_ref() {
-                    match self.peer_map.save_to_file(path).await {
+                    match self.inner.peer_map.save_to_file(path).await {
                         Ok(count) => {
                             debug!(count, "known peers persisted")
                         }
@@ -1171,10 +1740,6 @@ impl Actor {
                 self.port_mapper.deactivate();
                 self.derp_actor_sender
                     .send(DerpActorMessage::Shutdown)
-                    .await
-                    .ok();
-                self.udp_actor_sender
-                    .send(UdpActorMessage::Shutdown)
                     .await
                     .ok();
 
@@ -1188,9 +1753,6 @@ impl Actor {
 
                 debug!("shutdown complete");
                 return true;
-            }
-            ActorMessage::CloseOrReconnect(region_id, reason) => {
-                self.send_derp_actor(DerpActorMessage::CloseOrReconnect { region_id, reason });
             }
             ActorMessage::ReStun(reason) => {
                 self.re_stun(reason).await;
@@ -1209,14 +1771,6 @@ impl Actor {
                 self.set_preferred_port(port).await;
                 let _ = s.send(());
             }
-            ActorMessage::SendCallMeMaybe { dst, dst_key, msg } => {
-                let msg = disco::Message::CallMeMaybe(msg);
-                let _res = self.send_disco_message(dst, dst_key, msg).await;
-            }
-            ActorMessage::AddKnownAddr(addr, s) => {
-                self.add_known_addr(addr);
-                s.send(()).ok();
-            }
             ActorMessage::ReceiveDerp(read_result) => {
                 let passthroughs = self.process_derp_read_result(read_result).await;
                 for passthrough in passthroughs {
@@ -1230,42 +1784,14 @@ impl Actor {
                     }
                 }
             }
-            ActorMessage::EndpointPingExpired(id, txid) => {
-                if let Some(ep) = self.peer_map.by_id_mut(&id) {
+            ActorMessage::EndpointPingExpired(id, txid) => self.inner.peer_map.write(|pm| {
+                if let Some(ep) = pm.by_id_mut(&id).as_mut() {
                     ep.ping_timeout(txid);
                 }
-            }
-            ActorMessage::PublishAddrInfo => {
-                self.publish_addr_info();
-            }
+            }),
         }
 
         false
-    }
-
-    /// This modifies the [`quinn_udp::RecvMeta`] for the packet to set the addresses
-    /// to those that the QUIC layer should see.  E.g. the remote address will be set to the
-    /// [`QuicMappedAddr`] instead of the actual remote. It also registers activity for this peer.
-    ///
-    /// Returns `true` if the message should be processed.
-    fn receive_ip(&mut self, bytes: &Bytes, meta: &mut quinn_udp::RecvMeta) -> bool {
-        debug!("received data {} from {}", meta.len, meta.addr);
-        match self.peer_map.receive_ip(meta.addr) {
-            None => {
-                warn!(peer=?meta.addr, "no peer_map state found for peer, skipping");
-                return false;
-            }
-            Some(ep) => {
-                debug!("peer_map state found for {}", meta.addr);
-                meta.addr = ep.quic_mapped_addr.0;
-            }
-        }
-
-        // Normalize local_ip
-        meta.dst_ip = self.normalized_local_addr().ok().map(|addr| addr.ip());
-
-        debug!("received passthrough message {}", bytes.len());
-        true
     }
 
     fn normalized_local_addr(&self) -> io::Result<SocketAddr> {
@@ -1288,35 +1814,38 @@ impl Actor {
         (ipv4_addr, ipv6_addr)
     }
 
-    async fn process_derp_read_result(&mut self, dm: DerpReadResult) -> Vec<NetworkReadResult> {
+    async fn process_derp_read_result(&mut self, dm: DerpReadResult) -> Vec<DerpRecvResult> {
         debug!("process_derp_read {} bytes", dm.buf.len());
         if dm.buf.is_empty() {
             warn!("received empty derp packet");
             return Vec::new();
         }
         let region_id = dm.region_id;
-        let ipp = SendAddr::Derp(region_id);
 
-        let ep_quic_mapped_addr = match self.peer_map.endpoint_for_node_key_mut(&dm.src) {
-            Some(ep) => {
+        let ep_quic_mapped_addr = self.inner.peer_map.write(|pm| {
+            let ep_quic_mapped_addr = pm.endpoint_for_node_key_mut(&dm.src).as_mut().map(|ep| {
                 // NOTE: we don't update the derp region if there is already one but the new one is
                 // different
                 if ep.derp_region().is_none() {
                     ep.set_derp_region(region_id);
                 }
                 ep.quic_mapped_addr
+            });
+
+            match ep_quic_mapped_addr {
+                Some(addr) => addr,
+                None => {
+                    info!(peer=%dm.src, "no peer_map state found for peer");
+                    let id = pm.insert_endpoint(EndpointOptions {
+                        public_key: dm.src,
+                        derp_region: Some(region_id),
+                        active: true,
+                    });
+                    let ep = pm.by_id_mut(&id).expect("inserted");
+                    ep.quic_mapped_addr
+                }
             }
-            None => {
-                info!(peer=%dm.src, "no peer_map state found for peer");
-                let id = self.peer_map.insert_endpoint(EndpointOptions {
-                    public_key: dm.src,
-                    derp_region: Some(region_id),
-                    active: true,
-                });
-                let ep = self.peer_map.by_id_mut(&id).expect("inserted");
-                ep.quic_mapped_addr
-            }
-        };
+        });
 
         // the derp packet is made up of multiple udp packets, prefixed by a u16 be length prefix
         //
@@ -1329,9 +1858,9 @@ impl Actor {
         for part in parts {
             match part {
                 Ok(part) => {
-                    if self.handle_derp_disco_message(&part, ipp, dm.src).await {
+                    if self.handle_derp_disco_message(&part, region_id, dm.src) {
                         // Message was internal, do not bubble up.
-                        debug!("processed internal disco message from {:?}", dm.src);
+                        debug!(peer = %dm.src, "processed internal disco message");
                         continue;
                     }
 
@@ -1342,124 +1871,15 @@ impl Actor {
                         dst_ip,
                         ecn: None,
                     };
-                    out.push(NetworkReadResult::Ok {
-                        source: NetworkSource::Derp,
-                        bytes: part,
-                        meta,
-                    });
+                    out.push(Ok((meta, part)));
                 }
                 Err(e) => {
-                    out.push(NetworkReadResult::Error(e));
+                    out.push(Err(e));
                 }
             }
         }
 
         out
-    }
-
-    /// Split a number of transmits into individual packets.
-    ///
-    /// For each transmit, if it has a segment size, it will be split into
-    /// multiple packets according to that segment size. If it does not have a
-    /// segment size, the contents will be sent as a single packet.
-    fn split_packets(transmits: Vec<quinn_udp::Transmit>) -> Vec<Bytes> {
-        let mut res = Vec::with_capacity(transmits.len());
-        for transmit in transmits {
-            let contents = transmit.contents;
-            if let Some(segment_size) = transmit.segment_size {
-                for chunk in contents.chunks(segment_size) {
-                    res.push(contents.slice_ref(chunk));
-                }
-            } else {
-                res.push(contents);
-            }
-        }
-        res
-    }
-
-    async fn send_network(&mut self, transmits: Vec<quinn_udp::Transmit>) {
-        trace!(
-            "sending:\n{}",
-            transmits.iter().fold(
-                String::with_capacity(transmits.len() * 50),
-                |mut final_repr, t| {
-                    final_repr.push_str(
-                        format!(
-                            "  dest: {}, src: {:?}, content_len: {}\n",
-                            QuicMappedAddr(t.destination),
-                            t.src_ip,
-                            t.contents.len()
-                        )
-                        .as_str(),
-                    );
-                    final_repr
-                }
-            )
-        );
-
-        if transmits.is_empty() {
-            return;
-        }
-        let current_destination = &transmits[0].destination;
-        debug_assert!(
-            transmits
-                .iter()
-                .all(|t| &t.destination == current_destination),
-            "mixed destinations"
-        );
-        let current_destination = QuicMappedAddr(*current_destination);
-
-        match self
-            .peer_map
-            .endpoint_for_quic_mapped_addr_mut(&current_destination)
-        {
-            Some(ep) => {
-                let public_key = *ep.public_key();
-                trace!(
-                    "Sending to endpoint for {:?} ({:?})",
-                    current_destination,
-                    public_key
-                );
-
-                let (udp_addr, derp_region, msgs) = ep.get_send_addrs();
-                self.handle_ping_actions(msgs).await;
-                match (udp_addr, derp_region) {
-                    (Some(udp_addr), Some(derp_region)) => {
-                        let res = self.send_raw(udp_addr, transmits.clone()).await;
-                        self.send_derp(derp_region, public_key, Self::split_packets(transmits));
-
-                        if let Err(err) = res {
-                            warn!("failed to send UDP: {:?}", err);
-                        }
-                    }
-                    (None, Some(derp_region)) => {
-                        self.send_derp(derp_region, public_key, Self::split_packets(transmits));
-                    }
-                    (Some(udp_addr), None) => {
-                        if let Err(err) = self.send_raw(udp_addr, transmits).await {
-                            warn!("failed to send UDP: {:?}", err);
-                        }
-                    }
-                    (None, None) => {
-                        warn!("no UDP or DERP addr")
-                    }
-                }
-            }
-            None => {
-                // TODO: This should not be possible.  We should have errored during the
-                // AsyncUdpSocket::poll_send call.
-                error!(addr=%current_destination, "no endpoint for mapped address");
-            }
-        }
-    }
-
-    #[instrument(level = "debug", skip_all)]
-    fn send_derp(&mut self, region_id: u16, peer: PublicKey, contents: Vec<Bytes>) {
-        self.send_derp_actor(DerpActorMessage::Send {
-            region_id,
-            contents,
-            peer,
-        });
     }
 
     /// Triggers an address discovery. The provided why string is for debug logging only.
@@ -1484,16 +1904,6 @@ impl Actor {
         }
     }
 
-    fn publish_addr_info(&self) {
-        if let Some(ref discovery) = self.inner.discovery {
-            let info = AddrInfo {
-                direct_addresses: self.last_endpoints.iter().map(|x| x.addr).collect(),
-                derp_region: Some(self.inner.my_derp()).filter(|x| *x != 0),
-            };
-            discovery.publish(&info);
-        }
-    }
-
     #[instrument(level = "debug", skip_all)]
     async fn update_endpoints(&mut self, why: &'static str) {
         inc!(MagicsockMetrics, update_endpoints);
@@ -1515,7 +1925,6 @@ impl Actor {
                     if let Some(ref cb) = self.inner.on_endpoints {
                         cb(&endpoints[..]);
                     }
-                    self.publish_addr_info();
                 }
             }
             Err(err) => {
@@ -1888,15 +2297,17 @@ impl Actor {
 
     #[instrument(skip_all, fields(me = %self.inner.me))]
     async fn enqueue_call_me_maybe(&mut self, derp_region: u16, endpoint_id: usize) {
-        let endpoint = self.peer_map.by_id(&endpoint_id);
-        if endpoint.is_none() {
+        let public_key = self
+            .inner
+            .peer_map
+            .read(|pm| pm.by_id(&endpoint_id).map(|ep| ep.public_key));
+        let Some(public_key) = public_key else {
             warn!(
                 "enqueue_call_me_maybe with invalid endpoint_id called: {} - {}",
                 derp_region, endpoint_id
             );
             return;
-        }
-        let endpoint = endpoint.unwrap();
+        };
         if self.last_endpoints_time.is_none()
             || self.last_endpoints_time.as_ref().unwrap().elapsed()
                 > ENDPOINTS_FRESH_ENOUGH_DURATION
@@ -1929,24 +2340,13 @@ impl Actor {
                 .await
                 .unwrap();
         } else {
-            let public_key = *endpoint.public_key();
             let eps: Vec<_> = self.last_endpoints.iter().map(|ep| ep.addr).collect();
             let msg = disco::CallMeMaybe { my_number: eps };
-
-            let msg_sender = self.msg_sender.clone();
-            tokio::task::spawn(async move {
-                warn!("sending call me maybe to {public_key:?}");
-                if let Err(err) = msg_sender
-                    .send(ActorMessage::SendCallMeMaybe {
-                        dst: SendAddr::Derp(derp_region),
-                        dst_key: public_key,
-                        msg,
-                    })
-                    .await
-                {
-                    warn!("failed to send disco message to {}: {:?}", derp_region, err);
-                }
-            });
+            self.inner.send_disco_message_derp(
+                derp_region,
+                public_key,
+                disco::Message::CallMeMaybe(msg),
+            )
         }
     }
 
@@ -1967,9 +2367,7 @@ impl Actor {
     /// This is called when connectivity changes enough that we no longer trust the old routes.
     #[instrument(skip_all, fields(me = %self.inner.me))]
     fn reset_endpoint_states(&mut self) {
-        for (_, ep) in self.peer_map.endpoints_mut() {
-            ep.note_connectivity_change();
-        }
+        self.inner.peer_map.reset_endpoint_states()
     }
 
     /// Closes and re-binds the UDP sockets.
@@ -2040,370 +2438,31 @@ impl Actor {
         }
     }
 
-    #[instrument(skip_all)]
-    async fn send_disco_message(
-        &mut self,
-        dst: SendAddr,
-        dst_key: PublicKey,
-        msg: disco::Message,
-    ) -> Result<bool> {
-        debug!("sending disco message to {}: {:?}", dst, msg);
-        if self.inner.is_closed() {
-            bail!("connection closed");
-        }
-        let di = get_disco_info(&mut self.disco_info, &self.inner.secret_key, &dst_key);
-        let mut seal = msg.as_bytes();
-        di.shared_key.seal(&mut seal);
-
-        let is_derp = dst.is_derp();
-        if is_derp {
-            inc!(MagicsockMetrics, send_disco_derp);
-        } else {
-            inc!(MagicsockMetrics, send_disco_udp);
-        }
-
-        let pkt = disco::encode_message(&self.inner.public_key(), seal);
-        let sent = self.send_addr(dst, Some(&dst_key), pkt.into()).await;
-        match sent {
-            Ok(0) => {
-                // Can't send. (e.g. no IPv6 locally)
-                warn!("disco: failed to send {:?} to {}", msg, dst);
-                Ok(false)
-            }
-            Ok(_n) => {
-                debug!("disco: sent message to {}", dst);
-                if is_derp {
-                    inc!(MagicsockMetrics, sent_disco_derp);
-                } else {
-                    inc!(MagicsockMetrics, sent_disco_udp);
-                }
-                match msg {
-                    disco::Message::Ping(_) => {
-                        inc!(MagicsockMetrics, sent_disco_ping);
-                    }
-                    disco::Message::Pong(_) => {
-                        inc!(MagicsockMetrics, sent_disco_pong);
-                    }
-                    disco::Message::CallMeMaybe(_) => {
-                        inc!(MagicsockMetrics, sent_disco_call_me_maybe);
-                    }
-                }
-                Ok(true)
-            }
-            Err(err) => {
-                warn!("disco: failed to send {:?} to {}: {:?}", msg, dst, err);
-                Err(err.into())
-            }
-        }
-    }
-
-    /// Sends either to UDP or DERP, depending on the IP.
-    #[instrument(skip_all)]
-    async fn send_addr(
-        &mut self,
-        addr: SendAddr,
-        pub_key: Option<&PublicKey>,
-        pkt: Bytes,
-    ) -> io::Result<usize> {
-        match addr {
-            SendAddr::Udp(addr) => {
-                let transmits = vec![quinn_udp::Transmit {
-                    destination: addr,
-                    contents: pkt,
-                    ecn: None,
-                    segment_size: None,
-                    src_ip: None, // TODO
-                }];
-                self.send_raw(addr, transmits).await
-            }
-            SendAddr::Derp(region) => match pub_key {
-                None => Err(io::Error::new(
-                    io::ErrorKind::Other,
-                    "missing pub key for derp route",
-                )),
-                Some(pub_key) => {
-                    self.send_derp(region, *pub_key, vec![pkt]);
-                    Ok(1)
-                }
-            },
-        }
-    }
-
-    async fn handle_derp_disco_message(
+    fn handle_derp_disco_message(
         &mut self,
         msg: &[u8],
-        src: SendAddr,
+        region: u16,
         derp_node_src: PublicKey,
     ) -> bool {
         match disco::source_and_box(msg) {
             Some((source, sealed_box)) => {
-                self.handle_disco_message(source, sealed_box, src, Some(derp_node_src))
-                    .await
+                self.inner.handle_disco_message(
+                    source,
+                    sealed_box,
+                    DiscoMessageSource::Derp {
+                        region,
+                        key: derp_node_src,
+                    },
+                );
+                true
             }
             None => false,
         }
     }
 
-    /// Handles a discovery message and reports whether `msg`f was a Tailscale inter-node discovery message.
-    ///
-    /// For messages received over DERP, the src.ip() will be DERP_MAGIC_IP (with src.port() being the region ID) and the
-    /// derp_node_src will be the node key it was received from at the DERP layer. derp_node_src is None when received over UDP.
-    #[instrument(skip_all)]
-    async fn handle_disco_message(
-        &mut self,
-        sender: PublicKey,
-        sealed_box: &[u8],
-        src: SendAddr,
-        derp_node_src: Option<PublicKey>,
-    ) -> bool {
-        debug!("handle_disco_message start {} - {:?}", src, derp_node_src);
-        if self.inner.is_closed() {
-            return true;
-        }
-
-        let mut unknown_sender = false;
-        if self.peer_map.endpoint_for_node_key(&sender).is_none() {
-            unknown_sender = match src {
-                SendAddr::Udp(addr) => self.peer_map.endpoint_for_ip_port_mut(addr).is_none(),
-                SendAddr::Derp(_) => true,
-            };
-        }
-
-        if unknown_sender {
-            // Disco Ping from unseen endpoint. We will have to add the
-            // endpoint later if the message is a ping
-            info!("disco: unknown sender {:?} - {}", sender, src);
-        }
-
-        // We're now reasonably sure we're expecting communication from
-        // this peer, do the heavy crypto lifting to see what they want.
-
-        let di = get_disco_info(&mut self.disco_info, &self.inner.secret_key, &sender);
-        let mut sealed_box = sealed_box.to_vec();
-        let payload = di.shared_key.open(&mut sealed_box);
-        if payload.is_err() {
-            // This could happen if we changed the key between restarts.
-            warn!(
-                "disco: [{:?}] failed to open box from {:?} (wrong rcpt?) {:?}",
-                self.inner.public_key(),
-                sender,
-                payload,
-            );
-            inc!(MagicsockMetrics, recv_disco_bad_key);
-            return true;
-        }
-        let dm = disco::Message::from_bytes(&sealed_box);
-        debug!("disco: disco.parse = {:?}", dm);
-
-        if dm.is_err() {
-            // Couldn't parse it, but it was inside a correctly
-            // signed box, so just ignore it, assuming it's from a
-            // newer version of Tailscale that we don't
-            // understand. Not even worth logging about, lest it
-            // be too spammy for old clients.
-
-            inc!(MagicsockMetrics, recv_disco_bad_parse);
-            return true;
-        }
-
-        let dm = dm.unwrap();
-        let is_derp = src.is_derp();
-        if is_derp {
-            inc!(MagicsockMetrics, recv_disco_derp);
-        } else {
-            inc!(MagicsockMetrics, recv_disco_udp);
-        }
-
-        debug!("got disco message: {:?}", dm);
-        match dm {
-            disco::Message::Ping(ping) => {
-                inc!(MagicsockMetrics, recv_disco_ping);
-                // if we get here we got a valid ping from an unknown sender
-                // so insert an endpoint for them
-                if unknown_sender {
-                    warn!(
-                        "unknown sender: {:?} with region id {:?}",
-                        sender,
-                        src.derp_region()
-                    );
-                    self.peer_map.insert_endpoint(EndpointOptions {
-                        public_key: sender,
-                        derp_region: src.derp_region(),
-                        active: true,
-                    });
-                }
-                self.handle_ping(ping, &sender, src, derp_node_src).await;
-                true
-            }
-            disco::Message::Pong(pong) => {
-                inc!(MagicsockMetrics, recv_disco_pong);
-                if let Some(ep) = self.peer_map.endpoint_for_node_key_mut(&sender) {
-                    let insert = ep
-                        .handle_pong_conn(&self.inner.public_key(), &pong, di, src)
-                        .await;
-                    if let Some((src, key)) = insert {
-                        self.peer_map.set_node_key_for_ip_port(src, &key);
-                    }
-                }
-                true
-            }
-            disco::Message::CallMeMaybe(cm) => {
-                inc!(MagicsockMetrics, recv_disco_call_me_maybe);
-                if !is_derp || derp_node_src.is_none() {
-                    // CallMeMaybe messages should only come via DERP.
-                    debug!("[unexpected] CallMeMaybe packets should only come via DERP");
-                    return true;
-                }
-                let node_key = derp_node_src.unwrap();
-                match self.peer_map.endpoint_for_node_key_mut(&node_key) {
-                    None => {
-                        inc!(MagicsockMetrics, recv_disco_call_me_maybe_bad_disco);
-                        debug!(
-                            "disco: ignoring CallMeMaybe from {:?}; {:?} is unknown",
-                            sender, node_key,
-                        );
-                    }
-                    Some(ep) => {
-                        info!(
-                            "disco: {:?}<-{:?} ({:?})  got call-me-maybe, {} endpoints",
-                            self.inner.public_key(),
-                            ep.public_key(),
-                            src,
-                            cm.my_number.len()
-                        );
-                        ep.handle_call_me_maybe(cm).await;
-                    }
-                }
-                true
-            }
-        }
-    }
-
-    /// di is the DiscoInfo of the source of the ping.
-    /// derp_node_src is non-zero if the ping arrived via DERP.
-    #[instrument(skip_all)]
-    async fn handle_ping(
-        &mut self,
-        dm: disco::Ping,
-        sender: &PublicKey,
-        src: SendAddr,
-        derp_node_src: Option<PublicKey>,
-    ) {
-        let di = get_disco_info(&mut self.disco_info, &self.inner.secret_key, sender);
-        let likely_heart_beat = Some(src) == di.last_ping_from
-            && di
-                .last_ping_time
-                .map(|s| s.elapsed() < Duration::from_secs(5))
-                .unwrap_or_default();
-        di.last_ping_from.replace(src);
-        di.last_ping_time.replace(Instant::now());
-        // If we got a ping over DERP, then derp_node_src is non-zero and we reply
-        // over DERP (in which case ip_dst is also a DERP address).
-        // But if the ping was over UDP (ip_dst is not a DERP address), then dst_key
-        // will be zero here, but that's fine: send_disco_message only requires
-        // a dstKey if the dst ip:port is DERP.
-
-        let dst_key = match derp_node_src {
-            Some(dst_key) => {
-                if !src.is_derp() {
-                    error!(%src, from=%sender.fmt_short(), "ignoring ping reported both as direct and relayed");
-                    return debug_assert!(false, "`derp_node_src` is some but `src` is not derp");
-                }
-                dst_key
-            }
-            None => {
-                if src.is_derp() {
-                    error!(%src, from=%sender.fmt_short(), "ignoring ping reported both as direct and relayed");
-                    return debug_assert!(false, "`derp_node_src` is none but `src` is derp");
-                }
-                di.node_key
-            }
-        };
-
-        if let Some(ep) = self.peer_map.endpoint_for_node_key_mut(&dst_key) {
-            if ep.endpoint_confirmed(src, dm.tx_id) {
-                debug!("disco: ping got duplicate endpoint {} - {}", src, dm.tx_id);
-                return;
-            }
-            if let SendAddr::Udp(addr) = src {
-                self.peer_map.set_node_key_for_ip_port(addr, &dst_key);
-            }
-        }
-
-        if !likely_heart_beat {
-            info!(
-                "disco: {:?}<-{:?} ({dst_key:?}, {src:?})  got ping tx={:?}",
-                self.inner.public_key(),
-                di.node_key,
-                dm.tx_id
-            );
-        }
-
-        let ip_dst = src;
-        let pong = disco::Message::Pong(disco::Pong {
-            tx_id: dm.tx_id,
-            src: src.as_socket_addr(),
-        });
-        if let Err(err) = self.send_disco_message(ip_dst, dst_key, pong).await {
-            warn!("disco: failed to send message to {ip_dst}: {err:?}");
-        }
-    }
-
-    #[instrument(skip_all)]
-    fn add_known_addr(&mut self, peer_addr: PeerAddr) {
-        self.peer_map.add_peer_addr(peer_addr)
-    }
-
     /// Returns the current IPv4 listener's port number.
     fn local_port_v4(&self) -> u16 {
         self.pconn4.port()
-    }
-
-    #[instrument(skip_all)]
-    async fn send_raw(
-        &self,
-        addr: SocketAddr,
-        mut transmits: Vec<quinn_udp::Transmit>,
-    ) -> io::Result<usize> {
-        debug!("send_raw: {} packets", transmits.len());
-
-        if addr.is_ipv6() && self.pconn6.is_none() {
-            return Err(io::Error::new(io::ErrorKind::Other, "no IPv6 connection"));
-        }
-
-        let conn = if addr.is_ipv6() {
-            self.pconn6.as_ref().unwrap()
-        } else {
-            &self.pconn4
-        };
-
-        if transmits.iter().any(|t| t.destination != addr) {
-            for t in &mut transmits {
-                t.destination = addr;
-            }
-        }
-        let sum =
-            futures::future::poll_fn(|cx| conn.poll_send(&self.udp_state, cx, &transmits)).await?;
-        let total_bytes: u64 = transmits
-            .iter()
-            .take(sum)
-            .map(|x| x.contents.len() as u64)
-            .sum();
-        if addr.is_ipv6() {
-            inc_by!(MagicsockMetrics, send_ipv6, total_bytes);
-        } else {
-            inc_by!(MagicsockMetrics, send_ipv4, total_bytes);
-        }
-
-        debug!("sent {} packets to {}", sum, addr);
-        debug_assert!(
-            sum <= transmits.len(),
-            "too many msgs {} > {}",
-            sum,
-            transmits.len()
-        );
-
-        Ok(sum)
     }
 }
 
@@ -2487,13 +2546,6 @@ impl SendAddr {
         matches!(self, Self::Derp(_))
     }
 
-    fn derp_region(&self) -> Option<u16> {
-        match self {
-            Self::Derp(region) => Some(*region),
-            Self::Udp(_) => None,
-        }
-    }
-
     /// Returns the mapped version or the actual `SocketAddr`.
     fn as_socket_addr(&self) -> SocketAddr {
         match self {
@@ -2521,43 +2573,24 @@ impl Display for SendAddr {
     }
 }
 
-/// A simple iterator to group [`Transmit`]s by destination.
+/// Split a number of transmits into individual packets.
 ///
-/// [`Transmit`]: quinn_udp::Transmit
-struct TransmitIter<'a> {
-    transmits: &'a [quinn_udp::Transmit],
-    offset: usize,
-}
-
-impl<'a> TransmitIter<'a> {
-    fn new(transmits: &'a [quinn_udp::Transmit]) -> Self {
-        TransmitIter {
-            transmits,
-            offset: 0,
-        }
-    }
-}
-
-impl Iterator for TransmitIter<'_> {
-    type Item = Vec<quinn_udp::Transmit>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        if self.offset == self.transmits.len() {
-            return None;
-        }
-        let current_dest = &self.transmits[self.offset].destination;
-        let mut end = self.offset;
-        for t in &self.transmits[self.offset..] {
-            if current_dest != &t.destination {
-                break;
+/// For each transmit, if it has a segment size, it will be split into
+/// multiple packets according to that segment size. If it does not have a
+/// segment size, the contents will be sent as a single packet.
+fn split_packets(transmits: &[quinn_udp::Transmit]) -> DerpContents {
+    let mut res = SmallVec::with_capacity(transmits.len());
+    for transmit in transmits {
+        let contents = &transmit.contents;
+        if let Some(segment_size) = transmit.segment_size {
+            for chunk in contents.chunks(segment_size) {
+                res.push(contents.slice_ref(chunk));
             }
-            end += 1;
+        } else {
+            res.push(contents.clone());
         }
-
-        let out = self.transmits[self.offset..end].to_vec();
-        self.offset = end;
-        Some(out)
     }
+    res
 }
 
 /// Splits a packet into its component items.
@@ -2652,58 +2685,30 @@ impl std::fmt::Display for QuicMappedAddr {
         write!(f, "QuicMappedAddr({})", self.0)
     }
 }
+fn disco_message_sent(msg: &disco::Message) {
+    match msg {
+        disco::Message::Ping(_) => {
+            inc!(MagicsockMetrics, sent_disco_ping);
+        }
+        disco::Message::Pong(_) => {
+            inc!(MagicsockMetrics, sent_disco_pong);
+        }
+        disco::Message::CallMeMaybe(_) => {
+            inc!(MagicsockMetrics, sent_disco_call_me_maybe);
+        }
+    }
+}
 
 #[cfg(test)]
 pub(crate) mod tests {
     use anyhow::Context;
     use rand::RngCore;
-    use std::net::Ipv4Addr;
     use tokio::{net, sync, task::JoinSet};
     use tracing::{debug_span, Instrument};
     use tracing_subscriber::{prelude::*, EnvFilter};
 
     use super::*;
     use crate::{derp::DerpMode, test_utils::run_derper, tls, MagicEndpoint};
-
-    fn make_transmit(destination: SocketAddr) -> quinn_udp::Transmit {
-        quinn_udp::Transmit {
-            destination,
-            ecn: None,
-            contents: destination.to_string().into(),
-            segment_size: None,
-            src_ip: None,
-        }
-    }
-
-    #[test]
-    fn test_transmit_iter() {
-        let transmits = vec![
-            make_transmit(SocketAddr::new(Ipv4Addr::LOCALHOST.into(), 1)),
-            make_transmit(SocketAddr::new(Ipv4Addr::LOCALHOST.into(), 2)),
-            make_transmit(SocketAddr::new(Ipv4Addr::LOCALHOST.into(), 2)),
-            make_transmit(SocketAddr::new(Ipv4Addr::LOCALHOST.into(), 1)),
-            make_transmit(SocketAddr::new(Ipv4Addr::LOCALHOST.into(), 3)),
-            make_transmit(SocketAddr::new(Ipv4Addr::LOCALHOST.into(), 3)),
-        ];
-
-        let groups: Vec<_> = TransmitIter::new(&transmits).collect();
-        dbg!(&groups);
-        assert_eq!(groups.len(), 4);
-        assert_eq!(groups[0].len(), 1);
-        assert_eq!(groups[1].len(), 2);
-        assert_eq!(groups[2].len(), 1);
-        assert_eq!(groups[3].len(), 2);
-
-        let transmits = vec![
-            make_transmit(SocketAddr::new(Ipv4Addr::LOCALHOST.into(), 1)),
-            make_transmit(SocketAddr::new(Ipv4Addr::LOCALHOST.into(), 1)),
-        ];
-
-        let groups: Vec<_> = TransmitIter::new(&transmits).collect();
-        dbg!(&groups);
-        assert_eq!(groups.len(), 1);
-        assert_eq!(groups[0].len(), 2);
-    }
 
     async fn pick_port() -> u16 {
         let conn = net::UdpSocket::bind("127.0.0.1:0").await.unwrap();
@@ -2859,7 +2864,7 @@ pub(crate) mod tests {
                         direct_addresses: new_eps.iter().map(|ep| ep.addr).collect(),
                     },
                 };
-                let _ = m.endpoint.magic_sock().add_peer_addr(addr).await;
+                m.endpoint.magic_sock().add_peer_addr(addr);
             }
         }
 
@@ -3139,8 +3144,8 @@ pub(crate) mod tests {
             Ok(quic_ep)
         };
 
-        let m1 = make_conn("127.0.0.1:8770".parse().unwrap())?;
-        let m2 = make_conn("127.0.0.1:8771".parse().unwrap())?;
+        let m1 = make_conn("127.0.0.1:0".parse().unwrap())?;
+        let m2 = make_conn("127.0.0.1:0".parse().unwrap())?;
 
         // msg from  a -> b
         macro_rules! roundtrip {
@@ -3398,9 +3403,9 @@ pub(crate) mod tests {
     }
 
     #[test]
-    fn split_packets() {
+    fn test_split_packets() {
         fn mk_transmit(contents: &[u8], segment_size: Option<usize>) -> quinn_udp::Transmit {
-            let destination = "127.0.0.1:12345".parse().unwrap();
+            let destination = "127.0.0.1:0".parse().unwrap();
             quinn_udp::Transmit {
                 destination,
                 ecn: None,
@@ -3409,17 +3414,17 @@ pub(crate) mod tests {
                 src_ip: None,
             }
         }
-        fn mk_expected(parts: impl IntoIterator<Item = &'static str>) -> Vec<Bytes> {
+        fn mk_expected(parts: impl IntoIterator<Item = &'static str>) -> DerpContents {
             parts
                 .into_iter()
                 .map(|p| p.as_bytes().to_vec().into())
                 .collect()
         }
         // no packets
-        assert_eq!(Actor::split_packets(vec![]), Vec::<Bytes>::default());
+        assert_eq!(split_packets(&[]), SmallVec::<[Bytes; 1]>::default());
         // no split
         assert_eq!(
-            Actor::split_packets(vec![
+            split_packets(&vec![
                 mk_transmit(b"hello", None),
                 mk_transmit(b"world", None)
             ]),
@@ -3427,12 +3432,12 @@ pub(crate) mod tests {
         );
         // split without rest
         assert_eq!(
-            Actor::split_packets(vec![mk_transmit(b"helloworld", Some(5)),]),
+            split_packets(&[mk_transmit(b"helloworld", Some(5))]),
             mk_expected(["hello", "world"])
         );
         // split with rest and second transmit
         assert_eq!(
-            Actor::split_packets(vec![
+            split_packets(&vec![
                 mk_transmit(b"hello world", Some(5)),
                 mk_transmit(b"!", None)
             ]),
@@ -3440,7 +3445,7 @@ pub(crate) mod tests {
         );
         // split that results in 1 packet
         assert_eq!(
-            Actor::split_packets(vec![
+            split_packets(&vec![
                 mk_transmit(b"hello world", Some(1000)),
                 mk_transmit(b"!", None)
             ]),
