@@ -1,29 +1,30 @@
-use std::collections::BTreeMap;
 use std::str::FromStr;
 use std::{net::SocketAddr, path::PathBuf, time::Duration};
 
 use anyhow::Result;
 use bytes::Bytes;
 use clap::{Args, Parser, Subcommand};
+use colored::Colorize;
 use comfy_table::presets::NOTHING;
 use comfy_table::{Cell, Table};
 use console::style;
 use futures::{Stream, StreamExt};
 use human_time::ToHumanTimeString;
 use indicatif::{
-    HumanBytes, HumanDuration, ProgressBar, ProgressDrawTarget, ProgressState, ProgressStyle,
+    HumanBytes, HumanDuration, MultiProgress, ProgressBar, ProgressDrawTarget, ProgressState,
+    ProgressStyle,
 };
 use iroh::client::quic::Iroh;
 use iroh::dial::Ticket;
 use iroh::rpc_protocol::*;
 use iroh_bytes::{protocol::RequestToken, util::runtime, BlobFormat, Hash, Tag};
+use iroh_net::magicsock::DirectAddrInfo;
 use iroh_net::PeerAddr;
 use iroh_net::{
     key::{PublicKey, SecretKey},
     magic_endpoint::ConnectionInfo,
 };
 
-use crate::commands::sync::fmt_short;
 use crate::config::{ConsoleEnv, NodeConfig};
 
 use self::node::{RpcPort, StartOptions};
@@ -367,7 +368,16 @@ impl NodeCommands {
         match self {
             Self::Connections => {
                 let connections = iroh.node.connections().await?;
-                println!("{}", fmt_connections(connections).await);
+                let timestamp = time::OffsetDateTime::now_utc()
+                    .format(&time::format_description::well_known::Rfc2822)
+                    .unwrap_or_else(|_| String::from("failed to get current time"));
+
+                println!(
+                    " {}: {}\n\n{}",
+                    "current time".bold(),
+                    timestamp,
+                    fmt_connections(connections).await
+                );
             }
             Self::Connection { node_id } => {
                 let conn_info = iroh.node.connection_info(node_id).await?;
@@ -678,46 +688,120 @@ async fn fmt_connections(
 ) -> String {
     let mut table = Table::new();
     table.load_preset(NOTHING).set_header(
-        vec!["node id", "region", "conn type", "latency"]
+        ["node id", "region", "conn type", "latency", "last used"]
             .into_iter()
             .map(bold_cell),
     );
     while let Some(Ok(conn_info)) = infos.next().await {
-        let node_id = conn_info.public_key.to_string();
+        let node_id: Cell = conn_info.public_key.to_string().into();
         let region = conn_info
             .derp_region
-            .map_or(String::new(), |region| region.to_string());
-        let conn_type = conn_info.conn_type.to_string();
+            .map_or(String::new(), |region| region.to_string())
+            .into();
+        let conn_type = conn_info.conn_type.to_string().into();
         let latency = match conn_info.latency {
             Some(latency) => latency.to_human_time_string(),
             None => String::from("unknown"),
-        };
-        table.add_row(vec![node_id, region, conn_type, latency]);
+        }
+        .into();
+        let last_used = conn_info
+            .last_used
+            .map(fmt_how_long_ago)
+            .map(Cell::new)
+            .unwrap_or_else(never);
+        table.add_row([node_id, region, conn_type, latency, last_used]);
     }
     table.to_string()
 }
 
 fn fmt_connection(info: ConnectionInfo) -> String {
-    format!(
-        "node_id: {}\nderp_region: {}\nconnection type: {}\nlatency: {}\n\n{}",
-        fmt_short(info.public_key),
-        info.derp_region
-            .map_or(String::from("unknown"), |r| r.to_string()),
-        info.conn_type,
-        fmt_latency(info.latency),
-        fmt_addrs(info.addrs)
-    )
+    let ConnectionInfo {
+        id: _,
+        public_key,
+        derp_region,
+        addrs,
+        conn_type,
+        latency,
+        last_used,
+    } = info;
+    let timestamp = time::OffsetDateTime::now_utc()
+        .format(&time::format_description::well_known::Rfc2822)
+        .unwrap_or_else(|_| String::from("failed to get current time"));
+    let mut table = Table::new();
+    table.load_preset(NOTHING);
+    table.add_row([bold_cell("current time"), timestamp.into()]);
+    table.add_row([bold_cell("node id"), public_key.to_string().into()]);
+    let derp_region = derp_region
+        .map(|r| r.to_string())
+        .unwrap_or_else(|| String::from("unknown"));
+    table.add_row([bold_cell("derp region"), derp_region.into()]);
+    table.add_row([bold_cell("connection type"), conn_type.to_string().into()]);
+    table.add_row([bold_cell("latency"), fmt_latency(latency).into()]);
+    table.add_row([
+        bold_cell("last used"),
+        last_used
+            .map(fmt_how_long_ago)
+            .map(Cell::new)
+            .unwrap_or_else(never),
+    ]);
+    table.add_row([bold_cell("known addresses"), addrs.len().into()]);
+
+    let general_info = table.to_string();
+
+    let addrs_info = fmt_addrs(addrs);
+    format!("{general_info}\n\n{addrs_info}",)
 }
 
-fn fmt_addrs(addrs: Vec<(SocketAddr, Option<Duration>)>) -> String {
+fn direct_addr_row(info: DirectAddrInfo) -> comfy_table::Row {
+    let DirectAddrInfo {
+        addr,
+        latency,
+        last_control,
+        last_payload,
+    } = info;
+
+    let last_control = match last_control {
+        None => never(),
+        Some((how_long_ago, kind)) => {
+            format!("{kind} ( {} )", fmt_how_long_ago(how_long_ago)).into()
+        }
+    };
+    let last_payload = last_payload
+        .map(fmt_how_long_ago)
+        .map(Cell::new)
+        .unwrap_or_else(never);
+
+    [
+        addr.into(),
+        fmt_latency(latency).into(),
+        last_control,
+        last_payload,
+    ]
+    .into()
+}
+
+fn fmt_addrs(addrs: Vec<DirectAddrInfo>) -> comfy_table::Table {
     let mut table = Table::new();
+    table.load_preset(NOTHING).set_header(
+        vec!["addr", "latency", "last control", "last data"]
+            .into_iter()
+            .map(bold_cell),
+    );
+    table.add_rows(addrs.into_iter().map(direct_addr_row));
     table
-        .load_preset(NOTHING)
-        .set_header(vec!["addr", "latency"].into_iter().map(bold_cell));
-    for addr in addrs {
-        table.add_row(vec![addr.0.to_string(), fmt_latency(addr.1)]);
-    }
-    table.to_string()
+}
+
+fn never() -> Cell {
+    Cell::new("never").add_attribute(comfy_table::Attribute::Dim)
+}
+
+fn fmt_how_long_ago(duration: Duration) -> String {
+    duration
+        .to_human_time_string()
+        .split_once(',')
+        .map(|(first, _rest)| first)
+        .unwrap_or("-")
+        .to_string()
 }
 
 fn fmt_latency(latency: Option<Duration>) -> String {
@@ -727,15 +811,24 @@ fn fmt_latency(latency: Option<Duration>) -> String {
     }
 }
 
-const PROGRESS_STYLE: &str =
-    "{msg}\n{spinner:.green} [{elapsed_precise}] [{wide_bar:.cyan/blue}] {bytes}/{total_bytes} ({eta})";
-
-fn make_download_pb() -> ProgressBar {
+fn make_overall_progress() -> ProgressBar {
     let pb = ProgressBar::hidden();
-    pb.set_draw_target(ProgressDrawTarget::stderr());
-    pb.enable_steady_tick(std::time::Duration::from_millis(50));
+    pb.enable_steady_tick(std::time::Duration::from_millis(100));
     pb.set_style(
-        ProgressStyle::with_template(PROGRESS_STYLE)
+        ProgressStyle::with_template(
+            "{msg}{spinner:.green} [{elapsed_precise}] [{wide_bar:.cyan/blue}] {pos}/{len}",
+        )
+        .unwrap()
+        .progress_chars("#>-"),
+    );
+    pb
+}
+
+fn make_individual_progress() -> ProgressBar {
+    let pb = ProgressBar::hidden();
+    pb.enable_steady_tick(std::time::Duration::from_millis(100));
+    pb.set_style(
+        ProgressStyle::with_template("{msg}{spinner:.green} [{elapsed_precise}] [{wide_bar:.cyan/blue}] {bytes}/{total_bytes} ({eta})")
             .unwrap()
             .with_key(
                 "eta",
@@ -750,67 +843,51 @@ fn make_download_pb() -> ProgressBar {
 
 pub async fn show_download_progress(
     hash: Hash,
-    mut stream: impl Stream<Item = Result<GetProgress>> + Unpin,
+    mut stream: impl Stream<Item = Result<DownloadProgress>> + Unpin,
 ) -> Result<()> {
     eprintln!("Fetching: {}", hash);
-    let pb = make_download_pb();
-    pb.set_message(format!("{} Connecting ...", style("[1/3]").bold().dim()));
-    let mut sizes = BTreeMap::new();
-    let mut downloading = false;
+    let mp = MultiProgress::new();
+    mp.set_draw_target(ProgressDrawTarget::stderr());
+    let op = mp.add(make_overall_progress());
+    let ip = mp.add(make_individual_progress());
+    op.set_message(format!("{} Connecting ...\n", style("[1/3]").bold().dim()));
+    let mut seq = false;
     while let Some(x) = stream.next().await {
         match x? {
-            GetProgress::Connected => {
-                pb.set_message(format!("{} Requesting ...", style("[2/3]").bold().dim()));
+            DownloadProgress::Connected => {
+                op.set_message(format!("{} Requesting ...\n", style("[2/3]").bold().dim()));
             }
-            GetProgress::FoundCollection {
-                total_blobs_size,
-                num_blobs,
-                ..
-            } => {
-                let count = num_blobs.unwrap_or_default();
-                let missing_bytes = total_blobs_size.unwrap_or_default();
-                pb.set_message(format!(
-                    "{} Downloading {} file(s) with total transfer size {}",
+            DownloadProgress::FoundHashSeq { children, .. } => {
+                op.set_message(format!(
+                    "{} Downloading {} blob(s)\n",
                     style("[3/3]").bold().dim(),
-                    count,
-                    HumanBytes(missing_bytes),
+                    children + 1,
                 ));
-                pb.set_length(missing_bytes);
-                pb.reset();
-                downloading = true;
+                op.set_length(children + 1);
+                op.reset();
+                seq = true;
             }
-            GetProgress::Found { id, size, .. } => {
-                if !downloading {
-                    pb.set_message(format!(
-                        "{} Downloading blob with size {}",
-                        style("[3/3]").bold().dim(),
-                        size,
-                    ));
-                    pb.set_length(size);
-                    pb.reset();
+            DownloadProgress::Found { size, child, .. } => {
+                if seq {
+                    op.set_position(child);
+                } else {
+                    op.finish_and_clear();
                 }
-                sizes.insert(id, (size, 0));
+                ip.set_length(size);
+                ip.reset();
             }
-            GetProgress::Progress { id, offset } => {
-                if let Some((_, current)) = sizes.get_mut(&id) {
-                    *current = offset;
-                    let total = sizes.values().map(|(_, current)| current).sum::<u64>();
-                    pb.set_position(total);
-                }
+            DownloadProgress::Progress { offset, .. } => {
+                ip.set_position(offset);
             }
-            GetProgress::Done { id } => {
-                if let Some((size, current)) = sizes.get_mut(&id) {
-                    *current = *size;
-                    let total = sizes.values().map(|(_, current)| current).sum::<u64>();
-                    pb.set_position(total);
-                }
+            DownloadProgress::Done { .. } => {
+                ip.finish_and_clear();
             }
-            GetProgress::NetworkDone {
+            DownloadProgress::NetworkDone {
                 bytes_read,
                 elapsed,
                 ..
             } => {
-                pb.finish_and_clear();
+                op.finish_and_clear();
                 eprintln!(
                     "Transferred {} in {}, {}/s",
                     HumanBytes(bytes_read),
@@ -818,7 +895,7 @@ pub async fn show_download_progress(
                     HumanBytes((bytes_read as f64 / elapsed.as_secs_f64()) as u64)
                 );
             }
-            GetProgress::AllDone => {
+            DownloadProgress::AllDone => {
                 break;
             }
             _ => {}
